@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "clinica"
-version = "0.7.7"
+version = "0.8.0"
 description = "Software platform for clinical neuroimaging studies"
 license = "MIT"
 authors = ["ARAMIS Lab"]
@@ -27,22 +27,21 @@ classifiers = [
 ]
 
 [tool.poetry.dependencies]
-python = ">=3.8,<3.12"
-nibabel = "^3.2.2"
-nipype = ">=1.7.1,<1.8.3"
+python = ">=3.9,<3.13"
+nibabel = "^5"
+nipype = ">=1.8.6"
 argcomplete = "^1.9.4"
 pandas = "^1.4"
 jinja2 = "^3"
 xvfbwrapper = "*"
 numpy = "^1.17"
 scikit-learn = "^1.0"
-nilearn = {version = "^0.9.2", extras = ["plotting"]}
+nilearn = "^0.10.3"
 colorlog = "^5"
 xgboost = "*"
 scipy = "^1.7"
 matplotlib = "*"
-niflow-nipype1-workflows = "*"
-scikit-image = "^0.21"
+scikit-image = "*"
 pydicom = "*"
 networkx = "^2"
 click = "^8"
@@ -63,10 +62,12 @@ pydra-petpvc="^0.0.4"
 pydra-fsl = "^0.0.22"
 
 [tool.poetry.group.dev.dependencies]
-black = "*"
-isort = "*"
-pre-commit = "*"
-codespell = {extras = ["toml"], version = "^2.2.2"}
+pytest = "*"
+pytest-cov = "*"
+pytest-mock = "*"
+pytest-random-order = "*"
+pytest-timeout = "*"
+pytest-xdist = "*"
 
 [tool.poetry.group.docs]
 optional = true
@@ -76,16 +77,6 @@ mkdocs = "^1.1"
 mkdocs-material = ">=7.1.8"
 pymdown-extensions = "*"
 
-[tool.poetry.group.test.dependencies]
-pytest = "*"
-pytest-timeout = "*"
-pytest-xdist = "*"
-pytest-cov = "^3.0.0"
-pytest-mock = "*"
-pytest-random-order = "*"
-scikit-image = "^0.21"
-nibabel = "*"
-
 [tool.poetry.scripts]
 clinica = "clinica.cmdline:main"
 
@@ -93,35 +84,27 @@ clinica = "clinica.cmdline:main"
 requires = ["poetry-core>=1.0.0"]
 build-backend = "poetry.core.masonry.api"
 
-[tool.black]
+[tool.ruff]
+target-version = "py38"
 line-length = 88
-target-version = ['py36', 'py37', 'py38']
-include = '\.pyi?$'
-force-exclude = '''
-/(
-    \.eggs
-  | \.git
-  | \.hg
-  | \.mypy_cache
-  | \.tox
-  | \.venv
-  | \.pytest_cache
-  | _build
-  | buck-out
-  | build
-  | dist
-  | docs
-  | clinica/lib
-)/
-'''
 
-[tool.isort]
-profile = "black"
+[tool.ruff.lint]
+select = [
+    "E",
+    "W",
+    "I001",
+#    "PTH",
+]
+ignore = ["E203", "E501"]
 
-[tool.flake8]
-max-line-length = 88
-select = ["E", "W"]
-extend-ignore = ["E203", "E501", "W503"]
+[tool.ruff.lint.isort]
+known-first-party = ["clinica"]
+
+[tool.ruff.format]
+quote-style = "double"
+indent-style = "space"
+skip-magic-trailing-comma = false
+line-ending = "auto"
 
 [tool.codespell]
 summary = ''
