@@ -4,10 +4,12 @@ Subclasses are located in clinica/pipeline/<pipeline_name>/<pipeline_name>_pipel
 """
 
 import abc
-from typing import Tuple
+from pathlib import Path
+from typing import List, Optional, Tuple
 
 import click
-from nipype.pipeline.engine import Workflow
+from nipype.interfaces.utility import IdentityInterface
+from nipype.pipeline.engine import Node, Workflow
 
 
 def clinica_pipeline(func):
@@ -46,26 +48,26 @@ def postset(attribute, value):
     return postset_decorator
 
 
-def detect_cross_sectional_and_longitudinal_subjects(
-    all_subs: list, bids_dir: str
-) -> Tuple[list, list]:
+def _detect_cross_sectional_and_longitudinal_subjects(
+    all_subs: List[str], bids_dir: Path
+) -> Tuple[List[str], List[str]]:
     """This function detects whether a subject has a longitudinal or cross sectional structure.
     If it has a mixed structure, it will be considered cross sectionnal.
 
     Parameters
     ----------
-    all_subs: list
+    all_subs: list of str
         List containing all the names of the subjects.
 
-    bids_dir: str
-        Path to the bids directory.
+    bids_dir: Path
+        The path to the bids directory.
 
     Returns
     -------
-    cross_subj: list
+    cross_subj: list of str
         List of all the subject detected with a cross_sectional organisation.
 
-    long_subj: list
+    long_subj: list of str
         List of all the subject detected with a longitudinal organisation.
 
     Examples
@@ -75,7 +77,7 @@ def detect_cross_sectional_and_longitudinal_subjects(
     │   └── anat
     └── sub-02
         └── ses-M000
-    >>> detect_cross_sectional_and_longitudinal_subjects(["sub-01", "sub-02"], /Users/name.surname/BIDS)
+    >>> _detect_cross_sectional_and_longitudinal_subjects(["sub-01", "sub-02"], /Users/name.surname/BIDS)
     (["sub-01"], ["sub-02])
     """
     from os import listdir
@@ -93,6 +95,235 @@ def detect_cross_sectional_and_longitudinal_subjects(
         else:
             long_subj.append(sub)
     return cross_subj, long_subj
+
+
+SYMBOLS = {
+    "customary": ("B", "K", "M", "G", "T", "P", "E", "Z", "Y"),
+    "customary_ext": (
+        "byte",
+        "kilo",
+        "mega",
+        "giga",
+        "tera",
+        "peta",
+        "exa",
+        "zetta",
+        "iotta",
+    ),
+    "iec": ("Bi", "Ki", "Mi", "Gi", "Ti", "Pi", "Ei", "Zi", "Yi"),
+    "iec_ext": (
+        "byte",
+        "kibi",
+        "mebi",
+        "gibi",
+        "tebi",
+        "pebi",
+        "exbi",
+        "zebi",
+        "yobi",
+    ),
+}
+
+
+def _bytes2human(
+    n: int, format: str = "%(value).1f %(symbol)s", symbols: str = "customary"
+) -> str:
+    """Convert n bytes into a human readable string based on format.
+
+    Symbols can be either "customary", "customary_ext", "iec" or "iec_ext", see: http://goo.gl/kTQMs
+
+    License:
+    Bytes-to-human / human-to-bytes converter.
+    Based on: http://goo.gl/kTQMs
+    Author: Giampaolo Rodola' <g.rodola [AT] gmail [DOT] com>
+    License: MIT
+    """
+    n = int(n)
+    if n < 0:
+        raise ValueError("n < 0")
+    symbols = SYMBOLS[symbols]
+    prefix = {}
+    for i, s in enumerate(symbols[1:]):
+        prefix[s] = 1 << (i + 1) * 10
+    for symbol in reversed(symbols[1:]):
+        if n >= prefix[symbol]:
+            value = float(n) / prefix[symbol]
+            return format % locals()
+    return format % dict(symbol=symbols[0], value=n)
+
+
+def _human2bytes(s: str) -> int:
+    """Attempts to guess the string format based on default symbols
+    set and return the corresponding bytes as an integer.
+
+    When unable to recognize the format ValueError is raised.
+
+    License:
+    Bytes-to-human / human-to-bytes converter.
+    Based on: http://goo.gl/kTQMs
+    Author: Giampaolo Rodola' <g.rodola [AT] gmail [DOT] com>
+    License: MIT
+    """
+    init = s
+    num = ""
+    while s and s[0:1].isdigit() or s[0:1] == ".":
+        num += s[0]
+        s = s[1:]
+    num = float(num)
+    letter = s.strip()
+    for name, sset in SYMBOLS.items():
+        if letter in sset:
+            break
+    else:
+        if letter == "k":
+            # treat 'k' as an alias for 'K' as per: http://goo.gl/kTQMs
+            sset = SYMBOLS["customary"]
+            letter = letter.upper()
+        else:
+            raise ValueError("can't interpret %r" % init)
+    prefix = {sset[0]: 1}
+    for i, s in enumerate(sset[1:]):
+        prefix[s] = 1 << (i + 1) * 10
+    return int(num * prefix[letter])
+
+
+def _add_session_label(filename: str) -> str:
+    """Transform a cross-sectional filename into a longitudinal one.
+
+    Examples
+    --------
+    >>> _add_ses("sub-ADNI001_scans.tsv")
+    sub-ADNI001_ses-M000_scans.tsv
+
+    No modification done if filename already has a session:
+
+    >>> _add_ses("sub-023a_ses-M012_T1w.nii.gz")
+    sub-023a_ses-M012_T1w.nii.gz
+
+    Parameters
+    ----------
+    filename : str
+
+    Returns
+    -------
+    str :
+        filename with '_ses-M000_ added just after participant_id
+    """
+    import re
+
+    # If filename contains ses-..., returns the original filename
+    # Regex explication:
+    # ^ start of string
+    # ([a-zA-Z0-9]*) matches any number of characters from a to z,
+    #       A to Z, 0 to 9, and store it in group(1)
+    # (?!ses-[a-zA-Z0-9]) do not match if there is already a 'ses-'
+    # (.*) catches the rest of the string
+    m = re.search(r"(^sub-[a-zA-Z0-9]*)_(?!ses-[a-zA-Z0-9])(.*)", filename)
+    try:
+        return m.group(1) + "_ses-M000_" + m.group(2)
+    except AttributeError:
+        return filename
+
+
+def _copy2_add_session_label(src: Path, dst: Path) -> Path:
+    """Calls copy2 function from shutil, but modifies the filename of the
+    copied files if they match the regex template described in the
+    _add_session_label() function
+
+    Parameters
+    ----------
+    src : Path
+        The path to the file that needs to be copied.
+
+    dst : Path
+        The original destination for the copied file.
+
+    Returns
+    -------
+    Path :
+        copy2 with modified filename.
+    """
+    from shutil import copy2
+
+    return copy2(src, dst / _add_session_label(src.name))
+
+
+def _convert_cross_sectional(
+    bids_in: Path,
+    bids_out: Path,
+    cross_subjects: List[str],
+    long_subjects: List[str],
+) -> None:
+    """
+    This function converts a cross-sectional-bids dataset into a
+    longitudinal clinica-compliant dataset
+
+    Parameters
+    ----------
+    bids_in : Path
+        The path to the cross-sectional bids dataset you want to convert.
+
+    bids_out : Path
+        The path to the converted longitudinal bids dataset.
+
+    cross_subjects : list of str
+        List of subjects in cross-sectional form (they need some adjustment).
+
+    long_subjects : list of str
+        List of subjects in longitudinal form (they just need to be copied).
+    """
+    if not bids_out.exists():
+        bids_out.mkdir()
+    _convert(bids_in, bids_out, cross_subjects, cross_sectional=True)
+    _convert(bids_in, bids_out, long_subjects, cross_sectional=False)
+
+
+def _convert(
+    bids_in: Path, bids_out: Path, subjects: List[str], cross_sectional: bool
+) -> None:
+    for subject in subjects:
+        output_folder = (
+            bids_out / subject / "ses-M000" if cross_sectional else bids_out / subject
+        )
+        if not output_folder.exists():
+            output_folder.mkdir(parents=True)
+        for file_to_copy in [
+            f for f in (bids_in / subject).iterdir() if not f.name.startswith(".")
+        ]:
+            copy_func = _copy_cross_sectional if cross_sectional else _copy_longitudinal
+            copy_func(bids_in / subject / file_to_copy, output_folder)
+
+
+def _copy_cross_sectional(file_to_copy: Path, output_folder: Path) -> None:
+    from shutil import copy2, copytree
+
+    if not (output_folder / file_to_copy.name).exists():
+        if file_to_copy.is_dir():
+            copytree(
+                file_to_copy,
+                output_folder / file_to_copy.name,
+                copy_function=_copy2_add_session_label,
+            )
+        elif file_to_copy.is_file():
+            new_filename_wo_ses = _add_session_label(file_to_copy.name)
+            copy2(file_to_copy, output_folder / new_filename_wo_ses)
+
+
+def _copy_longitudinal(file_to_copy: Path, output_folder: Path) -> None:
+    from shutil import copy2, copytree
+
+    if not (output_folder / file_to_copy.name).exists():
+        if file_to_copy.is_dir():
+            copytree(file_to_copy, output_folder / file_to_copy.name)
+        elif file_to_copy.is_file():
+            copy2(file_to_copy, output_folder)
+
+
+def _check_cross_subj(cross_subj: list) -> None:
+    from clinica.utils.exceptions import ClinicaInconsistentDatasetError
+
+    if len(cross_subj) > 0:
+        raise ClinicaInconsistentDatasetError(cross_subj)
 
 
 class Pipeline(Workflow):
@@ -113,76 +344,117 @@ class Pipeline(Workflow):
                 |
         [ Data Output Stream ]
 
-    Attributes:
-        is_built (bool): Informs if the pipelines has been built or not.
-        parameters (dict): Parameters of the pipelines.
-        info (dict): Information presented in the associated `info.json` file.
-        input_node (:obj:`npe.Node`): Identity interface connecting inputs.
-        output_node (:obj:`npe.Node`): Identity interface connecting outputs.
-        bids_directory (str): Directory used to read the data from, in BIDS.
-        caps_directory (str): Directory used to read/write the data from/to,
-            in CAPS.
-        subjects (list): List of subjects defined in the `subjects.tsv` file.
-            # TODO(@jguillon): Check the subjects-sessions file name.
-        sessions (list): List of sessions defined in the `subjects.tsv` file.
-        tsv_file (str): Path to the subjects-sessions `.tsv` file.
-        info_file (str): Path to the associated `info.json` file.
+    Attributes
+    ----------
+    is_built : bool
+        Informs if the pipelines has been built or not.
+
+    parameters : dict
+        Parameters of the pipelines.
+
+    info : dict
+        Information presented in the associated `info.json` file.
+
+    input_node : :obj:`npe.Node`
+        Identity interface connecting inputs.
+
+    output_node : :obj:`npe.Node`
+        Identity interface connecting outputs.
+
+    bids_directory : Path
+        Directory used to read the data from, in BIDS.
+
+    caps_directory : Path
+        Directory used to read/write the data from/to in CAPS.
+
+    subjects : list
+        List of subjects defined in the `subjects.tsv` file.
+        TODO(@jguillon): Check the subjects-sessions file name.
+
+    sessions : list
+        List of sessions defined in the `subjects.tsv` file.
+
+    tsv_file : str
+        Path to the subjects-sessions `.tsv` file.
+
+    info_file : str
+        Path to the associated `info.json` file.
     """
 
     __metaclass__ = abc.ABCMeta
 
     def __init__(
         self,
-        bids_directory=None,
-        caps_directory=None,
-        tsv_file=None,
-        overwrite_caps=False,
-        base_dir=None,
-        parameters={},
-        name=None,
+        bids_directory: Optional[str] = None,
+        caps_directory: Optional[str] = None,
+        tsv_file: Optional[str] = None,
+        overwrite_caps: Optional[bool] = False,
+        base_dir: Optional[str] = None,
+        parameters: Optional[dict] = None,
+        name: Optional[str] = None,
     ):
         """Init a Pipeline object.
 
-        Args:
-            bids_directory (str, optional): Path to a BIDS directory. Defaults to None.
-            caps_directory (str, optional): Path to a CAPS directory. Defaults to None.
-            tsv_file (str, optional): Path to a subjects-sessions `.tsv` file. Defaults to None.
-            overwrite_caps (bool, optional): Overwrite or not output directory.. Defaults to False.
-            base_dir (str, optional): Working directory (attribute of Nipype::Workflow class). Defaults to None.
-            parameters (dict, optional): Pipeline parameters. Defaults to {}.
-            name (str, optional): Pipeline name. Defaults to None.
+        Parameters
+        ----------
+        bids_directory : str, optional
+            Path to a BIDS directory. Defaults to None.
 
-        Raises:
-            RuntimeError: [description]
+        caps_directory : str, optional
+            Path to a CAPS directory. Defaults to None.
+
+        tsv_file : str, optional
+            Path to a subjects-sessions `.tsv` file. Defaults to None.
+
+        overwrite_caps : bool, optional
+            Overwrite or not output directory. Defaults to False.
+
+        base_dir : str, optional
+            Working directory (attribute of Nipype::Workflow class). Defaults to None.
+
+        parameters : dict, optional
+            Pipeline parameters. Defaults to {}.
+
+        name : str, optional
+            Pipeline name. Defaults to None.
+
+        Raises
+        ------
+        RuntimeError: [description]
         """
         import inspect
         import os
+        from pathlib import Path
         from tempfile import mkdtemp
 
         from clinica.utils.inputs import check_bids_folder, check_caps_folder
         from clinica.utils.participant import get_subject_session_list
 
-        self._is_built = False
-        self._overwrite_caps = overwrite_caps
-        self._bids_directory = bids_directory
-        self._caps_directory = caps_directory
-        self._verbosity = "debug"
-        self._tsv_file = tsv_file
-        self._info_file = os.path.join(
-            os.path.dirname(os.path.abspath(inspect.getfile(self.__class__))),
-            "info.json",
+        self._is_built: bool = False
+        self._overwrite_caps: bool = overwrite_caps
+        self._bids_directory: Optional[Path] = (
+            Path(bids_directory).absolute() if bids_directory else None
         )
-        self._info = {}
+        self._caps_directory: Optional[Path] = (
+            Path(caps_directory).absolute() if caps_directory else None
+        )
+        self._verbosity: str = "debug"
+        self._tsv_file: Optional[Path] = Path(tsv_file).absolute() if tsv_file else None
+        self._info_file: Path = (
+            Path(os.path.dirname(os.path.abspath(inspect.getfile(self.__class__))))
+            / "info.json"
+        )
+        self._info: dict = {}
 
         if base_dir:
-            self.base_dir = base_dir
+            self.base_dir = Path(base_dir).absolute()
             self._base_dir_was_specified = True
         else:
-            self.base_dir = mkdtemp()
+            self.base_dir = Path(mkdtemp())
             self._base_dir_was_specified = False
 
         self._name = name or self.__class__.__name__
-        self._parameters = parameters
+        self._parameters = parameters or {}
 
         if not self._bids_directory:
             if not self._caps_directory:
@@ -198,14 +470,98 @@ class Pipeline(Workflow):
             check_bids_folder(self._bids_directory)
             input_dir = self._bids_directory
             is_bids_dir = True
-        self._sessions, self._subjects = get_subject_session_list(
-            input_dir, tsv_file, is_bids_dir, False, base_dir
+        self._subjects, self._sessions = get_subject_session_list(
+            input_dir,
+            subject_session_file=tsv_file,
+            is_bids_dir=is_bids_dir,
+            use_session_tsv=False,
+            tsv_dir=base_dir,
         )
+        self._input_node = None
+        self._output_node = None
+        self._init_nodes()
 
+    @property
+    def base_dir_was_specified(self) -> bool:
+        return self._base_dir_was_specified
+
+    @property
+    def is_built(self) -> bool:
+        return self._is_built
+
+    @is_built.setter
+    def is_built(self, value: bool):
+        self._is_built = value
+
+    @property
+    def overwrite_caps(self) -> bool:
+        return self._overwrite_caps
+
+    @property
+    def parameters(self) -> dict:
+        return self._parameters
+
+    @parameters.setter
+    def parameters(self, value: dict):
+        self._parameters = value
+        # Need to rebuild input, output and core nodes
+        self.is_built = False
         self.init_nodes()
 
+    @property
+    def info(self) -> dict:
+        return self._info
+
+    @info.setter
+    def info(self, value: dict):
+        self._info = value
+
+    @property
+    def input_node(self) -> Node:
+        return self._input_node
+
+    @property
+    def output_node(self) -> Node:
+        return self._output_node
+
+    @property
+    def bids_directory(self) -> Optional[Path]:
+        return self._bids_directory
+
+    @property
+    def caps_directory(self) -> Optional[Path]:
+        return self._caps_directory
+
+    @property
+    def subjects(self) -> List[str]:
+        return self._subjects
+
+    @subjects.setter
+    def subjects(self, value: List[str]):
+        self._subjects = value
+        self.is_built = False
+
+    @property
+    def sessions(self) -> List[str]:
+        return self._sessions
+
+    @sessions.setter
+    def sessions(self, value: List[str]):
+        self._sessions = value
+        self.is_built = False
+
+    @property
+    def tsv_file(self) -> Optional[Path]:
+        return self._tsv_file
+
+    @property
+    def info_file(self) -> Path:
+        return self._info_file
+
     @staticmethod
-    def get_processed_images(caps_directory, subjects, sessions):
+    def get_processed_images(
+        caps_directory: Path, subjects: List[str], sessions: List[str]
+    ) -> List[str]:
         """Extract processed image IDs in `caps_directory` based on `subjects`_`sessions`.
 
         Todo:
@@ -221,58 +577,57 @@ class Pipeline(Workflow):
             "Implementation on which image(s) failed will appear soon."
         )
 
-    def init_nodes(self):
+    def _init_nodes(self) -> None:
         """Init the basic workflow and I/O nodes necessary before build."""
-        import nipype.interfaces.utility as nutil
-        import nipype.pipeline.engine as npe
-
-        if self.get_input_fields():
-            self._input_node = npe.Node(
-                name="Input",
-                interface=nutil.IdentityInterface(
-                    fields=self.get_input_fields(), mandatory_inputs=False
-                ),
-            )
-        else:
-            self._input_node = None
-
-        if self.get_output_fields():
-            self._output_node = npe.Node(
-                name="Output",
-                interface=nutil.IdentityInterface(
-                    fields=self.get_output_fields(), mandatory_inputs=False
-                ),
-            )
-        else:
-            self._output_node = None
-
+        self._init_input_node()
+        self._init_output_node()
         Workflow.__init__(self, self._name, self.base_dir)
         if self.input_node:
             self.add_nodes([self.input_node])
         if self.output_node:
             self.add_nodes([self.output_node])
 
-    def has_input_connections(self):
+    def _init_input_node(self) -> None:
+        if self.get_input_fields():
+            self._input_node = Node(
+                name="Input",
+                interface=IdentityInterface(
+                    fields=self.get_input_fields(), mandatory_inputs=False
+                ),
+            )
+
+    def _init_output_node(self) -> None:
+        if self.get_output_fields():
+            self._output_node = Node(
+                name="Output",
+                interface=IdentityInterface(
+                    fields=self.get_output_fields(), mandatory_inputs=False
+                ),
+            )
+
+    def has_input_connections(self) -> bool:
         """Checks if the Pipeline's input node has been connected.
 
-        Returns:
+        Returns
+        -------
+        bool :
             True if the input node is connected, False otherwise.
         """
         if self.input_node:
             return self._graph.in_degree(self.input_node) > 0
-        else:
-            return False
+        return False
 
-    def has_output_connections(self):
+    def has_output_connections(self) -> bool:
         """Checks if the Pipeline's output node has been connected.
 
-        Returns:
+        Returns
+        -------
+        bool :
             True if the output node is connected, False otherwise.
         """
         if self.output_node:
             return self._graph.out_degree(self.output_node) > 0
-        else:
-            return False
+        return False
 
     @postset("is_built", True)
     def build(self):
@@ -285,20 +640,27 @@ class Pipeline(Workflow):
         Since this method returns the concerned object, it can be chained to
         any other method of the Pipeline class.
 
-        Returns:
-            self: A Pipeline object.
+        Returns
+        -------
+        self: A Pipeline object.
         """
         if not self.is_built:
-            self.check_dependencies()
-            self.check_pipeline_parameters()
+            self._check_dependencies()
+            self._check_pipeline_parameters()
             if not self.has_input_connections():
-                self.build_input_node()
-            self.build_core_nodes()
+                self._build_input_node()
+            self._build_core_nodes()
             if not self.has_output_connections():
-                self.build_output_node()
+                self._build_output_node()
         return self
 
-    def run(self, plugin=None, plugin_args=None, update_hash=False, bypass_check=False):
+    def run(
+        self,
+        plugin=None,
+        plugin_args=None,
+        update_hash: bool = False,
+        bypass_check: bool = False,
+    ):
         """Executes the Pipeline.
 
         It overwrites the default Workflow method to check if the
@@ -323,11 +685,11 @@ class Pipeline(Workflow):
 
         if not self.is_built:
             self.build()
-        self.check_not_cross_sectional()
+        self._check_not_cross_sectional()
 
         if not bypass_check:
-            self.check_size()
-            plugin_args = self.update_parallelize_info(plugin_args)
+            self._check_size()
+            plugin_args = self._update_parallelize_info(plugin_args)
             plugin = "MultiProc"
         exec_graph = []
         try:
@@ -362,7 +724,7 @@ class Pipeline(Workflow):
             exec_graph = Graph()
         return exec_graph
 
-    def load_info(self):
+    def _load_info(self):
         """Loads the associated info.json file.
 
         Todo:
@@ -380,7 +742,7 @@ class Pipeline(Workflow):
             self.info = json.load(info_file)
         return self
 
-    def check_dependencies(self):
+    def _check_dependencies(self):
         """Checks if listed dependencies are present.
 
         Loads the pipelines related `info.json` file and check each one of the
@@ -401,32 +763,13 @@ class Pipeline(Workflow):
         Returns:
             self: A Pipeline object.
         """
-        import clinica.utils.check_dependency as chk
+        from clinica.utils.check_dependency import check_binary, check_software
 
-        # Checking functions preparation
-        check_software = {
-            # 'matlab': chk.check_matlab,
-            "ants": chk.check_ants,
-            "convert3d": chk.check_convert3d,
-            "spm": chk.check_spm,
-            "freesurfer": chk.check_freesurfer,
-            "fsl": chk.check_fsl,
-            "mrtrix": chk.check_mrtrix,
-            "matlab": chk.check_matlab,
-            "petpvc": chk.check_petpvc,
-        }
-        check_binary = chk.is_binary_present
-        # check_toolbox = chk.is_toolbox_present
-        # check_pipeline = chk.is_pipeline_present
-
-        # Load the info.json file
         if not self.info:
-            self.load_info()
-
-        # Dependencies checking
+            self._load_info()
         for d in self.info["dependencies"]:
             if d["type"] == "software":
-                check_software[d["name"]]()
+                check_software(d["name"])
             elif d["type"] == "binary":
                 check_binary(d["name"])
             elif d["type"] == "toolbox":
@@ -437,111 +780,18 @@ class Pipeline(Workflow):
                 raise Exception(
                     f"Pipeline.check_dependencies() Unknown dependency type: '{d['type']}'."
                 )
-
-        self.check_custom_dependencies()
+        self._check_custom_dependencies()
 
         return self
 
-    def check_size(self):
+    def _check_size(self):
         """Check if the pipeline has enough space on the disk for both working directory and CAPS."""
-        import select
         import sys
         from os import statvfs
         from os.path import dirname
 
         from clinica.utils.stream import cprint
 
-        SYMBOLS = {
-            "customary": ("B", "K", "M", "G", "T", "P", "E", "Z", "Y"),
-            "customary_ext": (
-                "byte",
-                "kilo",
-                "mega",
-                "giga",
-                "tera",
-                "peta",
-                "exa",
-                "zetta",
-                "iotta",
-            ),
-            "iec": ("Bi", "Ki", "Mi", "Gi", "Ti", "Pi", "Ei", "Zi", "Yi"),
-            "iec_ext": (
-                "byte",
-                "kibi",
-                "mebi",
-                "gibi",
-                "tebi",
-                "pebi",
-                "exbi",
-                "zebi",
-                "yobi",
-            ),
-        }
-
-        def bytes2human(n, format="%(value).1f %(symbol)s", symbols="customary"):
-            """
-            Convert n bytes into a human readable string based on format.
-            Symbols can be either "customary", "customary_ext", "iec" or "iec_ext",
-            see: http://goo.gl/kTQMs
-
-            License:
-            Bytes-to-human / human-to-bytes converter.
-            Based on: http://goo.gl/kTQMs
-            Working with Python 2.x and 3.x.
-            Author: Giampaolo Rodola' <g.rodola [AT] gmail [DOT] com>
-            License: MIT
-            """
-            n = int(n)
-            if n < 0:
-                raise ValueError("n < 0")
-            symbols = SYMBOLS[symbols]
-            prefix = {}
-            for i, s in enumerate(symbols[1:]):
-                prefix[s] = 1 << (i + 1) * 10
-            for symbol in reversed(symbols[1:]):
-                if n >= prefix[symbol]:
-                    value = float(n) / prefix[symbol]
-                    return format % locals()
-            return format % dict(symbol=symbols[0], value=n)
-
-        def human2bytes(s):
-            """
-            Attempts to guess the string format based on default symbols
-            set and return the corresponding bytes as an integer.
-            When unable to recognize the format ValueError is raised.
-
-            License:
-            Bytes-to-human / human-to-bytes converter.
-            Based on: http://goo.gl/kTQMs
-            Working with Python 2.x and 3.x.
-            Author: Giampaolo Rodola' <g.rodola [AT] gmail [DOT] com>
-            License: MIT
-            """
-            init = s
-            num = ""
-            while s and s[0:1].isdigit() or s[0:1] == ".":
-                num += s[0]
-                s = s[1:]
-            num = float(num)
-            letter = s.strip()
-            for name, sset in SYMBOLS.items():
-                if letter in sset:
-                    break
-            else:
-                if letter == "k":
-                    # treat 'k' as an alias for 'K' as per: http://goo.gl/kTQMs
-                    sset = SYMBOLS["customary"]
-                    letter = letter.upper()
-                else:
-                    raise ValueError("can't interpret %r" % init)
-            prefix = {sset[0]: 1}
-            for i, s in enumerate(sset[1:]):
-                prefix[s] = 1 << (i + 1) * 10
-            return int(num * prefix[letter])
-
-        timeout = 15
-
-        # Get the number of sessions
         n_sessions = len(self.subjects)
         try:
             caps_stat = statvfs(self.caps_directory)
@@ -557,32 +807,32 @@ class Pipeline(Workflow):
         try:
             space_needed_caps_1_session = self.info["space_caps"]
             space_needed_wd_1_session = self.info["space_wd"]
-            space_needed_caps = n_sessions * human2bytes(space_needed_caps_1_session)
-            space_needed_wd = n_sessions * human2bytes(space_needed_wd_1_session)
+            space_needed_caps = n_sessions * _human2bytes(space_needed_caps_1_session)
+            space_needed_wd = n_sessions * _human2bytes(space_needed_wd_1_session)
             error = ""
             if free_space_caps == free_space_wd:
                 if space_needed_caps + space_needed_wd > free_space_wd:
                     # We assume this is the same disk
                     error = (
                         f"{error}Space needed for CAPS and working directory ("
-                        f"{bytes2human(space_needed_caps + space_needed_wd)}"
+                        f"{_bytes2human(space_needed_caps + space_needed_wd)}"
                         f") is greater than what is left on your hard drive ("
-                        f"{bytes2human(free_space_wd)})"
+                        f"{_bytes2human(free_space_wd)})"
                     )
             else:
                 if space_needed_caps > free_space_caps:
                     error = (
                         f"{error} Space needed for CAPS ("
-                        f"{bytes2human(space_needed_caps)}"
+                        f"{_bytes2human(space_needed_caps)}"
                         f") is greater than what is left on your hard drive ("
-                        f"{bytes2human(free_space_caps)})\n"
+                        f"{_bytes2human(free_space_caps)})\n"
                     )
                 if space_needed_wd > free_space_wd:
                     error = (
                         f"{error}Space needed for working_directory ("
-                        f"{bytes2human(space_needed_wd)}"
+                        f"{_bytes2human(space_needed_wd)}"
                         f") is greater than what is left on your hard drive ("
-                        f"{bytes2human(free_space_wd)})\n"
+                        f"{_bytes2human(free_space_wd)})\n"
                     )
             if error:
                 cprint(msg=f"[SpaceError] {error}", lvl="error")
@@ -595,26 +845,18 @@ class Pipeline(Workflow):
         except KeyError:
             cprint("No info on how much size the pipeline takes. Running anyway...")
 
-    def update_parallelize_info(self, plugin_args):
+    def _update_parallelize_info(self, plugin_args: Optional[dict]) -> dict:
         """Performs some checks of the number of threads given in parameters,
         given the number of CPUs of the machine in which clinica is running.
         We force the use of plugin MultiProc
 
-        Author: Arnaud Marcoux"""
-        import select
-        import sys
+        Author: Arnaud Marcoux
+        """
         from multiprocessing import cpu_count
 
         from clinica.utils.stream import cprint
 
-        # count number of CPUs
         n_cpu = cpu_count()
-        # timeout value: max time allowed to decide how many thread
-        # to run in parallel (sec)
-        timeout = 15
-
-        # Use this var to know in the end if we need to ask the user
-        # an other number
         ask_user = False
 
         try:
@@ -661,270 +903,54 @@ class Pipeline(Workflow):
 
         return plugin_args
 
-    def check_not_cross_sectional(self):
+    def _check_not_cross_sectional(self):
         """
-        This function checks if the dataset is longitudinal. If it is cross
-        sectional, clinica proposes to convert it in a clinica compliant form.
+        This function checks if the dataset is longitudinal.
+
+        If it is cross-sectional, clinica proposes to convert it in a clinica compliant form.
 
         author: Arnaud Marcoux
         """
         import sys
-        from os import listdir
-        from os.path import abspath, basename, dirname, isdir, join
 
         from clinica.utils.exceptions import ClinicaInconsistentDatasetError
         from clinica.utils.stream import cprint
 
-        def _check_cross_subj(cross_subj: list) -> None:
-            if len(cross_subj) > 0:
-                raise ClinicaInconsistentDatasetError(cross_subj)
-
-        def convert_cross_sectional(bids_in, bids_out, cross_subjects, long_subjects):
-            """
-            This function converts a cross-sectional-bids dataset into a
-            longitudinal clinica-compliant dataset
-
-            Args:
-                bids_in: cross sectional bids dataset you want to convert
-                bids_out: converted longitudinal bids dataset
-                cross_subjects: list of subjects in cross sectional form
-                (they need some adjustment)
-                long_subjects: list of subjects in longitudinal form (they
-                just need to be copied)
-
-            Returns:
-                nothing
-            """
-            from os import mkdir
-            from os.path import exists, isfile
-            from shutil import copy2, copytree
-
-            def add_ses(f):
-                """
-                Use this function to transform a cross sectional filename into
-                a longitudinal one.
-                Examples:
-                sub-ADNI001_scans.tsv -> sub-ADNI001_ses-M000_scans.tsv
-                sub-023a_ses-M012_T1w.nii.gz -> sub-023a_ses-M012_T1w.nii.gz (no
-                    modification done if filename already has a session)
-
-                Args:
-                    f: filename
-
-                Returns:
-                    filename with '_ses-M000_ added just after participant_id
-                """
-                import re
-
-                # If filename contains ses-..., returns the original filename
-                # Regex explication:
-                # ^ start of string
-                # ([a-zA-Z0-9]*) matches any number of characters from a to z,
-                #       A to Z, 0 to 9, and store it in group(1)
-                # (?!ses-[a-zA-Z0-9]) do not match if there is already a 'ses-'
-                # (.*) catches the rest of the string
-                m = re.search(r"(^sub-[a-zA-Z0-9]*)_(?!ses-[a-zA-Z0-9])(.*)", f)
-                try:
-                    return m.group(1) + "_ses-M000_" + m.group(2)
-                except AttributeError:
-                    # If something goes wrong, we return the original filename
-                    return f
-
-            def copy2_add_ses(src, dst):
-                """
-                copy2_add_ses calls copy2 function from shutil, but modifies
-                the filename of the copied files if they match the regex
-                template described in add_ses() function
-
-                Args:
-                    src: path to the file that needs to be copied
-                    dst: original destination for the copied file
-
-                Returns:
-                    copy2 with modified filename
-                """
-                from os.path import basename, dirname, join
-                from shutil import copy2
-
-                dst_modified = join(dirname(dst), add_ses(basename(src)))
-                return copy2(src, dst_modified)
-
-            if not exists(bids_out):
-                # Create the output folder if it does not exists yet
-                mkdir(bids_out)
-
-            # First part of the algorithm: deal with subjects that does not
-            # have longitudinal (session) information
-            for subj in cross_subjects:
-                # Get list of of files/folders to copy. Remove hidden element
-                # ( if they start with a dot '.')
-                to_copy = [
-                    f for f in listdir(join(bids_in, subj)) if not f.startswith(".")
-                ]
-                # Always check that folder are existing, otherwise an exception
-                # is raised even though that does not prevent the function from
-                # working
-                if not exists(join(bids_out, subj)):
-                    mkdir(join(bids_out, subj))
-                if not exists(join(bids_out, subj, "ses-M000")):
-                    mkdir(join(bids_out, subj, "ses-M000"))
-                for el in to_copy:
-                    path_el = join(bids_in, subj, el)
-                    if not exists(join(bids_out, subj, "ses-M000", el)):
-                        # If the element to copy is a folder...
-                        if isdir(path_el):
-                            # Copytree is used with a 'custom' copy function,
-                            # to give a correct filemename to the files inside
-                            # the copied folders
-                            copytree(
-                                path_el,
-                                join(bids_out, subj, "ses-M000", basename(path_el)),
-                                copy_function=copy2_add_ses,
-                            )
-                        # If the element to copy is a file...
-                        elif isfile(path_el):
-                            # Modify the filename with add_ses function
-                            new_filename_wo_ses = add_ses(el)
-                            copy2(
-                                path_el,
-                                join(bids_out, subj, "ses-M000", new_filename_wo_ses),
-                            )
-            # Second part of the algorithm: deal with subjects that do not
-            # have the problem. We only xopy the content of the folder, and no
-            # filename needs to be changed
-            for su in long_subjects:
-                # Do not copy hidden files
-                to_copy = [
-                    f for f in listdir(join(bids_in, su)) if not f.startswith(".")
-                ]
-                if not exists(join(bids_out, su)):
-                    mkdir(join(bids_out, su))
-                for el in to_copy:
-                    path_el = join(bids_in, su, el)
-                    if not exists(join(bids_out, su, el)):
-                        # 2 possible cases: element to pcopy is a folder or a
-                        # file
-                        if isdir(path_el):
-                            copytree(path_el, join(bids_out, su, basename(path_el)))
-                        elif isfile(path_el):
-                            copy2(path_el, join(bids_out, su))
-
-        if self.bids_directory is not None:
-            bids_dir = abspath(self.bids_directory)
-            # Extract all subjects in BIDS directory: element must be a folder and
-            # a name starting with 'sub-'
-            all_subs = [
-                f
-                for f in listdir(bids_dir)
-                if isdir(join(bids_dir, f)) and f.startswith("sub-")
-            ]
-
-            cross_subj, long_subj = detect_cross_sectional_and_longitudinal_subjects(
-                all_subs, bids_dir
+        if self.bids_directory is None:
+            return
+        all_subs = [
+            f.name
+            for f in self.bids_directory.iterdir()
+            if (self.bids_directory / f).is_dir() and f.name.startswith("sub-")
+        ]
+        cross_subj, long_subj = _detect_cross_sectional_and_longitudinal_subjects(
+            all_subs, self.bids_directory
+        )
+        try:
+            _check_cross_subj(cross_subj)
+        except ClinicaInconsistentDatasetError as e:
+            cprint(e, lvl="warning")
+            proposed_bids = (
+                self.bids_directory / f"{self.bids_directory.name}_clinica_compliant"
             )
-            try:
-                _check_cross_subj(cross_subj)
-            except ClinicaInconsistentDatasetError as e:
-                cprint(e, lvl="warning")
-                proposed_bids = join(
-                    dirname(bids_dir), basename(bids_dir) + "_clinica_compliant"
+            if not click.confirm(
+                "Do you want to proceed with the conversion in another folder? "
+                "(Your original BIDS folder will not be modified "
+                f"and the folder {proposed_bids} will be created.)"
+            ):
+                click.echo("Clinica will now exit...")
+                sys.exit()
+            else:
+                cprint("Converting cross-sectional dataset into longitudinal...")
+                _convert_cross_sectional(
+                    self.bids_directory, proposed_bids, cross_subj, long_subj
+                )
+                cprint(
+                    f"Conversion succeeded. Your clinica-compliant dataset is located here: {proposed_bids}"
                 )
 
-                if not click.confirm(
-                    "Do you want to proceed with the conversion in another "
-                    "folder? (Your original BIDS folder will not be modified "
-                    f"and the folder {proposed_bids} will be created.)"
-                ):
-                    click.echo("Clinica will now exit...")
-                    sys.exit()
-                else:
-                    cprint("Converting cross-sectional dataset into longitudinal...")
-                    convert_cross_sectional(
-                        bids_dir, proposed_bids, cross_subj, long_subj
-                    )
-                    cprint(
-                        f"Conversion succeeded. Your clinica-compliant dataset is located here: {proposed_bids}"
-                    )
-
-    @property
-    def base_dir_was_specified(self):
-        return self._base_dir_was_specified
-
-    @property
-    def is_built(self):
-        return self._is_built
-
-    @is_built.setter
-    def is_built(self, value):
-        self._is_built = value
-
-    @property
-    def overwrite_caps(self):
-        return self._overwrite_caps
-
-    @property
-    def parameters(self):
-        return self._parameters
-
-    @parameters.setter
-    def parameters(self, value):
-        self._parameters = value
-        # Need to rebuild input, output and core nodes
-        self.is_built = False
-        self.init_nodes()
-
-    @property
-    def info(self):
-        return self._info
-
-    @info.setter
-    def info(self, value):
-        self._info = value
-
-    @property
-    def input_node(self):
-        return self._input_node
-
-    @property
-    def output_node(self):
-        return self._output_node
-
-    @property
-    def bids_directory(self):
-        return self._bids_directory
-
-    @property
-    def caps_directory(self):
-        return self._caps_directory
-
-    @property
-    def subjects(self):
-        return self._subjects
-
-    @subjects.setter
-    def subjects(self, value):
-        self._subjects = value
-        self.is_built = False
-
-    @property
-    def sessions(self):
-        return self._sessions
-
-    @sessions.setter
-    def sessions(self, value):
-        self._sessions = value
-        self.is_built = False
-
-    @property
-    def tsv_file(self):
-        return self._tsv_file
-
-    @property
-    def info_file(self):
-        return self._info_file
-
     @abc.abstractmethod
-    def build_core_nodes(self):
+    def _build_core_nodes(self):
         """Builds the Pipeline's core nodes.
 
         This method should use and connect to the `Pipeline.input_node` one
@@ -933,7 +959,7 @@ class Pipeline(Workflow):
         """
 
     @abc.abstractmethod
-    def build_input_node(self):
+    def _build_input_node(self):
         """Builds the Pipeline's input data stream node.
 
         Warnings:
@@ -942,7 +968,7 @@ class Pipeline(Workflow):
         """
 
     @abc.abstractmethod
-    def build_output_node(self):
+    def _build_output_node(self):
         """Builds the Pipeline's output data stream node.
 
         Warnings:
@@ -951,7 +977,7 @@ class Pipeline(Workflow):
         """
 
     @abc.abstractmethod
-    def get_input_fields(self):
+    def get_input_fields(self) -> List[str]:
         """Lists the input fields of the Pipeline.
 
         Returns:
@@ -960,7 +986,7 @@ class Pipeline(Workflow):
         """
 
     @abc.abstractmethod
-    def get_output_fields(self):
+    def get_output_fields(self) -> List[str]:
         """Lists the output fields of the Pipeline.
 
         Returns:
@@ -969,16 +995,16 @@ class Pipeline(Workflow):
         """
 
     @abc.abstractmethod
-    def check_custom_dependencies(self):
+    def _check_custom_dependencies(self) -> None:
         """Check dependencies provided by the developer."""
 
     @abc.abstractmethod
-    def check_pipeline_parameters(self):
+    def _check_pipeline_parameters(self) -> None:
         """Check pipeline parameters."""
 
 
 class PETPipeline(Pipeline):
-    def check_pipeline_parameters(self):
+    def _check_pipeline_parameters(self) -> None:
         """Check pipeline parameters."""
         if "acq_label" not in self.parameters.keys():
             raise KeyError("Missing compulsory acq_label key in pipeline parameter.")
