@@ -1,7 +1,8 @@
-import os
+from pathlib import Path
 
 import pytest
 
+from clinica.pipelines.dwi.dti.utils import DTIBasedMeasure
 from clinica.utils.pet import ReconstructionMethod, Tracer
 
 
@@ -43,7 +44,7 @@ def test_bids_pet_nii_empty():
     from clinica.utils.input_files import bids_pet_nii
 
     assert bids_pet_nii() == {
-        "pattern": os.path.join("pet", f"*_pet.nii*"),
+        "pattern": Path("pet") / "*_pet.nii*",
         "description": "PET data",
     }
 
@@ -51,9 +52,8 @@ def test_bids_pet_nii_empty():
 @pytest.fixture
 def expected_bids_pet_query(tracer, reconstruction):
     return {
-        "pattern": os.path.join(
-            "pet", f"*_trc-{tracer.value}_rec-{reconstruction.value}_pet.nii*"
-        ),
+        "pattern": Path("pet")
+        / f"*_trc-{tracer.value}_rec-{reconstruction.value}_pet.nii*",
         "description": f"PET data with {tracer.value} tracer and reconstruction method {reconstruction.value}",
     }
 
@@ -64,3 +64,26 @@ def test_bids_pet_nii(tracer, reconstruction, expected_bids_pet_query):
     from clinica.utils.input_files import bids_pet_nii
 
     assert bids_pet_nii(tracer, reconstruction) == expected_bids_pet_query
+
+
+@pytest.mark.parametrize("dti_measure", DTIBasedMeasure)
+@pytest.mark.parametrize("space", [None, "*", "T1w"])
+def test_dwi_dti_query(dti_measure, space):
+    from clinica.utils.input_files import dwi_dti
+
+    space = space or "*"
+    assert dwi_dti(dti_measure, space=space) == {
+        "pattern": f"dwi/dti_based_processing/*/*_space-{space}_{dti_measure.value}.nii.gz",
+        "description": f"DTI-based {dti_measure.value} in space {space}.",
+        "needed_pipeline": "dwi_dti",
+    }
+
+
+def test_dwi_dti_query_error():
+    from clinica.utils.input_files import dwi_dti
+
+    with pytest.raises(
+        ValueError,
+        match="'foo' is not a valid DTIBasedMeasure",
+    ):
+        dwi_dti("foo")
