@@ -410,7 +410,7 @@ def create_scans_dict(
                 # Some flutemeta lines contain a non-coded string value at the second-to-last position. This value
                 # contains a comma which adds an extra column and shifts the remaining values to the right. In this
                 # case, we just remove the erroneous content and replace it with -4 which AIBL uses as n/a value.
-                on_bad_lines = (
+                on_bad_lines = (  # noqa: E731
                     lambda bad_line: bad_line[:-3] + [-4, bad_line[-1]]
                     if "flutemeta" in file_path and study_name == "AIBL"
                     else "error"
@@ -841,10 +841,10 @@ def _build_dcm2niix_command(
     bids_sidecar: bool = True,
 ) -> list:
     """Generate the dcm2niix command from user inputs."""
-    command = ["sudo", "dcm2niix", "-w", "0", "-f", output_fmt, "-o", output_dir]
+    command = ["dcm2niix", "-w", "0", "-f", output_fmt, "-o", str(output_dir)]
     command += ["-9", "-z", "y"] if compress else ["-z", "n"]
     command += ["-b", "y", "-ba", "y"] if bids_sidecar else ["-b", "n"]
-    command += [input_dir]
+    command += [str(input_dir)]
 
     return command
 
@@ -897,12 +897,16 @@ def run_dcm2niix(
         cprint(
             msg=(
                 "DICOM to BIDS conversion with dcm2niix failed:\n"
-                f"command: {command}\n"
-                #f"{completed_process.stdout.decode('utf-8')}"
+                f"command: {' '.join(command)}\n"
+                f"{completed_process.stdout.decode('utf-8')}"
             ),
             lvl="warning",
         )
         return False
+    cprint(
+        f"DICOM to BIDS conversion with dcm2niix failed:\ncommand: {' '.join(command)}\n",
+        lvl="debug",
+    )
     return True
 
 
