@@ -34,11 +34,10 @@ def convert_adni_flair(
         If True, pre-existing images in the BIDS directory
         will be erased and extracted again.
     """
-    from os import path
-
-    import pandas as pd
-
-    from clinica.iotools.converters.adni_to_bids.adni_utils import paths_to_bids
+    from clinica.iotools.converters.adni_to_bids.adni_utils import (
+        load_clinical_csv,
+        paths_to_bids,
+    )
     from clinica.utils.stream import cprint
 
     if not subjects:
@@ -48,10 +47,11 @@ def convert_adni_flair(
         subjects = list(adni_merge.PTID.unique())
 
     cprint(
-        f"Calculating paths of FLAIR images. Output will be stored in {conversion_dir}."
+        f"Calculating paths of FLAIR images. Output will be stored in {conversion_dir}.",
+        lvl="info",
     )
     images = compute_flair_paths(source_dir, csv_dir, subjects, conversion_dir)
-    cprint("Paths of FLAIR images found. Exporting images into BIDS ...")
+    cprint("Paths of FLAIR images found. Exporting images into BIDS ...", lvl="info")
     # flair_paths_to_bids(images, dest_dir)
     paths_to_bids(images, destination_dir, "flair", mod_to_update=mod_to_update)
     cprint(msg="FLAIR conversion done.", lvl="debug")
