@@ -1,4 +1,5 @@
 from enum import Enum, auto
+from pathlib import Path
 from typing import List, Optional, Tuple, Union
 
 import pandas as pd
@@ -73,7 +74,8 @@ def visits_to_timepoints(
                 visits[key_preferred_visit] = preferred_visit_name
             elif visits[key_preferred_visit] != preferred_visit_name:
                 cprint(
-                    f"[{modality}] Subject {subject} has multiple visits for one timepoint."
+                    f"[{modality}] Subject {subject} has multiple visits for one timepoint.",
+                    lvl="info",
                 )
 
             unique_visits.remove(preferred_visit_name)
@@ -94,7 +96,7 @@ def visits_to_timepoints(
             subject,
         )
         if closest_visit is None:
-            cprint(image, lvl="debug")
+            cprint(f"No closest visit found for image {image}", lvl="debug")
             continue
         key_min_visit = (
             closest_visit.VISCODE,
@@ -106,7 +108,8 @@ def visits_to_timepoints(
             visits[key_min_visit] = image[visit_field]
         elif visits[key_min_visit] != image[visit_field]:
             cprint(
-                f"[{modality}] Subject {subject} has multiple visits for one timepoint."
+                f"[{modality}] Subject {subject} has multiple visits for one timepoint.",
+                lvl="debug",
             )
 
     return visits
@@ -224,10 +227,11 @@ def _get_closest_visit(
     from clinica.utils.stream import cprint
 
     if len(visits) == 0:
-        # cprint(
-        #     "No corresponding timepoint in ADNIMERGE for "
-        #     f"subject {subject} in visit {image_visit}"
-        # )
+        cprint(
+            "No corresponding timepoint in ADNIMERGE for "
+            f"subject {subject} in visit {image_visit}",
+            lvl="info",
+        )
         return None
 
     if len(visits) == 1:
@@ -542,7 +546,8 @@ def get_images_pet(
 
         if original_pet_meta.empty:
             cprint(
-                f"No {modality} images metadata for subject {subject} and visit {qc_visit[viscode_field]}"
+                f"No {modality} images metadata for subject {subject} and visit {qc_visit[viscode_field]}",
+                lvl="info",
             )
             continue
 
@@ -638,10 +643,6 @@ def correct_diagnosis_sc_adni3(clinical_data_dir, participants_df):
     Returns:
         Corrected participants_df.
     """
-    from os import path
-
-    import pandas as pd
-
     from clinica.utils.stream import cprint
 
     diagnosis_dict = {1: "CN", 2: "MCI", 3: "AD"}
@@ -658,7 +659,7 @@ def correct_diagnosis_sc_adni3(clinical_data_dir, participants_df):
             participants_df.loc[alternative_id, "diagnosis_sc"] = diagnosis_sc
         except KeyError:
             cprint(
-                msg=(f"Unknown screening diagnosis for subject {alternative_id}."),
+                msg=f"Unknown screening diagnosis for subject {alternative_id}.",
                 lvl="warning",
             )
             participants_df.loc[alternative_id, "diagnosis_sc"] = "n/a"
@@ -972,23 +973,24 @@ def create_adni_sessions_dict(
                 # Filter rows with invalid session IDs.
                 df_filtered.dropna(subset="session_id", inplace=True)
 
-                if location == "ADNIMERGE.csv":
-                    df_filtered["AGE"] = df_filtered.apply(
-                        lambda x: update_age(x), axis=1
-                    )
-                df_subj_session = update_sessions_df(
-                    df_subj_session, df_filtered, df_sessions, location
-                )
-            else:
-                dict_column_correspondence = dict(
-                    zip(df_sessions["ADNI"], df_sessions["BIDS CLINICA"])
-                )
-                df_filtered.rename(columns=dict_column_correspondence, inplace=True)
-                df_filtered = df_filtered.loc[
-                    :, (~df_filtered.columns.isin(df_subj_session.columns))
-                ]
-                df_subj_session = pd.concat([df_subj_session, df_filtered], axis=1)
-
+            if location == "ADNIMERGE.csv":
+                df_filtered["AGE"] = df_filtered.apply(lambda x: update_age(x), axis=1)
+            df_subj_session = update_sessions_df(
+                df_subj_session, df_filtered, df_sessions, location
+            )
+        else:
+            cprint(
+                f"Clinical dataframe extracted from {location} is empty after filtering.",
+                lvl="debug",
+            )
+            dict_column_correspondence = dict(
+                zip(df_sessions["ADNI"], df_sessions["BIDS CLINICA"])
+            )
+            df_filtered.rename(columns=dict_column_correspondence, inplace=True)
+            df_filtered = df_filtered.loc[
+                :, (~df_filtered.columns.isin(df_subj_session.columns))
+            ]
+            df_subj_session = pd.concat([df_subj_session, df_filtered], axis=1)
     if df_subj_session.empty:
         raise ValueError("Empty dataset detected. Clinical data cannot be extracted.")
 
@@ -1250,19 +1252,45 @@ def find_image_path(images, source_dir, modality, prefix, id_field):
     return images
 
 
-def paths_to_bids(images, bids_dir, modality, mod_to_update=False):
+def paths_to_bids(
+    images: pd.DataFrame,
+    bids_dir: str,
+    modality: str,
+    mod_to_update: bool = False,
+    n_procs: Optional[int] = 1,
+) -> List[Path]:
     """Images in the list are converted and copied to directory in BIDS format.
 
-    Args:
-        images: List of images metadata and paths
-        bids_dir: Path to the output BIDS directory
-        modality: Imaging modality
-        mod_to_update: If True, pre-existing images in the BIDS directory will be erased and extracted agai n.
+    Parameters
+    ----------
+    images : pd.DataFrame
+        List of images metadata and paths.
+
+    bids_dir : str
+        Path to the output BIDS directory.
+
+    modality : str
+        Imaging modality.
+
+    mod_to_update : bool
+        If True, pre-existing images in the BIDS directory will be erased
+        and extracted again.
+
+    n_procs : int, optional
+        The number of processes to use for conversion.
+        Default=1.
+
+    Returns
+    -------
+    output_file_treated : list of paths
+        List of path to image files created.
+        This list contains None values for files where the
+        conversion wasn't successful.
     """
     from functools import partial
     from multiprocessing import Pool, cpu_count
 
-    if modality.lower() not in [
+    if modality.lower() not in (
         "t1",
         "dwi",
         "flair",
@@ -1273,12 +1301,12 @@ def paths_to_bids(images, bids_dir, modality, mod_to_update=False):
         "pib",
         "av45_fbb",
         "tau",
-    ]:
-        # This should never be reached
+    ):
         raise RuntimeError(
-            modality.lower() + " is not supported for conversion in paths_to_bids"
+            f"{modality.lower()} is not supported for conversion in paths_to_bids."
         )
 
+    bids_dir = Path(bids_dir)
     images_list = list([data for _, data in images.iterrows()])
     with Pool(processes=max(cpu_count() - 1, 1)) as pool:
         create_file_ = partial(
@@ -1292,29 +1320,38 @@ def paths_to_bids(images, bids_dir, modality, mod_to_update=False):
     return output_file_treated
 
 
-def create_file(image, modality, bids_dir, mod_to_update):
-    """
-    Image file is created at the corresponding output folder
-    as result of image conversion (DICOM to NIFTI) and centering,
-    as specified for each modality
+def create_file(
+    image: pd.Series, modality: str, bids_dir: Path, mod_to_update: bool
+) -> Optional[Path]:
+    """Creates an image file at the corresponding output folder.
 
-    Args:
-        image: Image metadata
-        modality: Imaging modality
-        bids_dir: Path to the output BIDS directory
-        mod_to_update: If True, pre-existing images in the BIDS directory will be erased and extracted again.
+    The image file is the result of image conversion (DICOM to NIFTI)
+    and centering, as specified for each modality
 
-    Returns:
-        [Returns]
+    Parameters
+    ----------
+    image : pd.Series
+        The image metadata.
+
+    modality : str
+        Imaging modality.
+
+    bids_dir : Path
+        The path to the output BIDS directory.
+
+    mod_to_update : bool
+        If True, pre-existing images in the BIDS directory will be
+        erased and extracted again.
+
+    Returns
+    -------
+    output_image : Path
+        The path to the output image created.
+        If the conversion wasn't successful, this function returns None.
     """
     import os
     import re
     import shutil
-    from glob import glob
-    from os import path
-    import time
-
-    from numpy import nan
 
     from clinica.iotools.bids_utils import run_dcm2niix
     from clinica.iotools.converter_utils import viscode_to_session
@@ -1355,37 +1392,37 @@ def create_file(image, modality, bids_dir, mod_to_update):
         },
         "fdg": {
             "output_path": "pet",
-            "output_filename": f"_trc-{Tracer.FDG}_rec-{ReconstructionMethod.CO_REGISTERED_AVERAGED}_pet",
+            "output_filename": f"_trc-{Tracer.FDG.value}_rec-{ReconstructionMethod.CO_REGISTERED_AVERAGED.value}_pet",
             "to_center": True,
             "json": "n",
         },
         "fdg_uniform": {
             "output_path": "pet",
-            "output_filename": f"_trc-{Tracer.FDG}_rec-{ReconstructionMethod.COREGISTERED_ISOTROPIC}_pet",
+            "output_filename": f"_trc-{Tracer.FDG.value}_rec-{ReconstructionMethod.COREGISTERED_ISOTROPIC.value}_pet",
             "to_center": True,
             "json": "n",
         },
         "pib": {
             "output_path": "pet",
-            "output_filename": f"_trc-{Tracer.PIB}_pet",
+            "output_filename": f"_trc-{Tracer.PIB.value}_pet",
             "to_center": True,
             "json": "n",
         },
         "av45": {
             "output_path": "pet",
-            "output_filename": f"_trc-{Tracer.AV45}_pet",
+            "output_filename": f"_trc-{Tracer.AV45.value}_pet",
             "to_center": True,
             "json": "n",
         },
         "fbb": {
             "output_path": "pet",
-            "output_filename": f"_trc-{Tracer.FBB}_pet",
+            "output_filename": f"_trc-{Tracer.FBB.value}_pet",
             "to_center": True,
             "json": "n",
         },
         "tau": {
             "output_path": "pet",
-            "output_filename": f"_trc-{Tracer.AV1451}_pet",
+            "output_filename": f"_trc-{Tracer.AV1451.value}_pet",
             "to_center": True,
             "json": "n",
         },
@@ -1398,110 +1435,47 @@ def create_file(image, modality, bids_dir, mod_to_update):
         modality = image.Tracer.lower()
 
     if not image.Path:
-        # Removed this, it clutters the output making it difficult to see what is converted - MT 03/29/24
-        # cprint(
-        #     msg=(
-        #         f"[{modality.upper()}] No path specified for {subject} "
-        #         f"in session {viscode}"
-        #     ),
-        #     lvl="info",
-        # )
-        return nan
-    else:
         cprint(
-            msg=(
-                f"[{modality.upper()}] Processing subject {subject}"
-                f" in session {viscode}"
-            ),
+            f"[{modality.upper()}] No path specified for {subject} in session {viscode}",
             lvl="info",
         )
+        return None
+    cprint(
+        f"[{modality.upper()}] Processing subject {subject} in session {viscode}",
+        lvl="info",
+    )
 
     session = viscode_to_session(viscode)
 
     image_path = image.Path
     image_id = image.Image_ID
-    fmap_flag = 0
-
-    # if modality == "fmap":
-    #     fmap_path = image_path
-    #     replacements = {
-    #         "Axial_rsfMRI__Eyes_Open_": "ADNI_gre_field_mapping",
-    #         "Extended_Resting_State_fMRI": "ADNI_gre_field_mapping",
-    #         "Axial_fcMRI__EYES_OPEN_": "ADNI_gre_field_mapping",
-    #         "Axial_fcMRI__Eyes_Open_": "ADNI_gre_field_mapping",
-    #         "Resting_State_fMRI": "ADNI_gre_field_mapping",
-    #         "Axial_fcMRI": "ADNI_gre_field_mapping",
-    #         "Axial_rsfMRI__Eyes_Open_": "ADNI_gre_field_mapping",
-    #         "Axial_rsfMRI__EYES_OPEN_": "ADNI_gre_field_mapping",
-    #         "Axial_rsfMRI__Eyes_Open__-phase_P_to_A": "ADNI_gre_field_mapping",
-    #         "ADNI_gre_field_mapping_-phase_P_to_A": "ADNI_gre_field_mapping",
-    #         "ADNI_gre_field_mapping_0_angle__EYES_OPEN_": "ADNI_gre_field_mapping",
-    #         "Axial_fcMRI__EYES_OPEN_": "ADNI_gre_field_mapping"
-    #     }
-    #     fmap_path = image_path
-    #     for original, replacement in replacements.items():
-    #         fmap_path = fmap_path.replace(original, replacement)
-    #     fmap_path = fmap_path[:-9] # Remove the image identifier just check if the visit exists
-    #     if os.path.exists(fmap_path):
-    #         cprint("[FMAP] Found corresponding field maps for " + subject + " at " + fmap_path)
-    #         fmap_flag = 1
-    #     else:
-    #         fmap_path = fmap_path.replace("ADNI_gre_field_mapping", "Field_mapping")
-    #         if os.path.exists(fmap_path):
-    #             cprint("[FMAP] Found corresponding field maps for " + subject + " at " + fmap_path)
-    #             fmap_flag = 1
-    #         else:
-    #             fmap_path = fmap_path.replace("Field_mapping", "Field_Mapping")
-    #             if os.path.exists(fmap_path):
-    #                 cprint("[FMAP] Found corresponding field maps for " + subject + " at " + fmap_path)
-    #                 fmap_flag = 1
-    #             else:
-    #                 fmap_path = fmap_path.replace("Field_Mapping", "Field_Mapping_phase_R-L")
-    #                 if os.path.exists(fmap_path):
-    #                     cprint("[FMAP] Found corresponding field maps for " + subject + " at " + fmap_path)
-    #                     fmap_flag = 1
-    #                 else:
-    #                     fmap_path = fmap_path.replace("Field_Mapping_phase_R-L", "Axial_Field_Mapping")
-    #                     if os.path.exists(fmap_path):
-    #                         cprint("[FMAP] Found corresponding field maps for " + subject + " at " + fmap_path)
-    #                         fmap_flag = 1
-    #                     else:
-    #                         fmap_path = image_path
-    #                         fmap_flag = 0
-
-    # If the original image is a DICOM, check if contains two DICOM
-    # inside the same folder
+    # If the original image is a DICOM, check if contains two DICOM inside the same folder
     if image.Is_Dicom:
-        image_path = check_two_dcm_folder(image_path, bids_dir, image_id)
+        image_path = check_two_dcm_folder(image_path, str(bids_dir), image_id)
     bids_subj = subject.replace("_", "")
-    output_path = path.join(
-        bids_dir,
-        "sub-ADNI" + bids_subj,
-        session,
-        modality_specific[modality]["output_path"],
+    output_path = (
+        bids_dir
+        / f"sub-ADNI{bids_subj}"
+        / session
+        / modality_specific[modality]["output_path"]
     )
     output_filename = (
-        "sub-ADNI"
-        + bids_subj
-        + "_"
-        + session
-        + modality_specific[modality]["output_filename"]
+        f"sub-ADNI{bids_subj}_{session}{modality_specific[modality]['output_filename']}"
     )
+    output_path.mkdir(parents=True, exist_ok=True)
 
     # If updated mode is selected, check if an old image is existing and remove it
-    existing_im = glob(path.join(output_path, output_filename + "*"))
-    if mod_to_update and len(existing_im) > 0:
-        print("Removing the old image...")
-        for im in existing_im:
-            os.remove(im)
-
-    if mod_to_update or not len(existing_im) > 0:
-
-        try:
-            os.makedirs(output_path)
-        except OSError:
-            # Folder already created with previous instance
-            pass
+    for image_to_remove in output_path.glob(f"{output_filename}*"):
+        if not mod_to_update:
+            cprint(
+                f"There exists already an image at {image_to_remove}. "
+                "The parameter 'mod_to_update' is set to False such that "
+                "images cannot be overwritten.",
+                lvl="warning",
+            )
+            return None
+        cprint(f"Removing old image {image_to_remove}...", lvl="info")
+        image_to_remove.unlink()
 
         generate_json = modality_specific[modality]["json"]
         if modality_specific[modality]["to_center"]:
@@ -1539,70 +1513,108 @@ def create_file(image, modality, bids_dir, mod_to_update):
                         trigger_time[res.start() : res.end()], "."
                     )
                     os.rename(trigger_time, no_trigger_time)
+    generate_json = modality_specific[modality]["json"]
+    zip_image = "n" if modality_specific[modality]["to_center"] else "y"
+    file_without_extension = output_path / output_filename
+    output_image = file_without_extension.with_suffix(".nii.gz")
 
-                # Removing ADC images if one is generated
-                adc_image = path.join(output_path, output_filename + "_ADC.nii.gz")
-                if os.path.exists(adc_image):
-                    os.remove(adc_image)
+    if image.Is_Dicom:
+        success = run_dcm2niix(
+            input_dir=image_path,
+            output_dir=output_path,
+            output_fmt=output_filename,
+            compress=True if zip_image == "y" else False,
+            bids_sidecar=True if generate_json == "y" else False,
+        )
+        if not success:
+            cprint(
+                f"Error converting image {image_path} for subject {subject} and session {session}",
+                lvl="warning",
+            )
+        # If "_t" - the trigger delay time - exists in dcm2niix output filename, we remove it
+        for trigger_time in output_path.glob(f"{output_filename}_t[0-9]*"):
+            res = re.search(r"_t\d+\.", str(trigger_time))
+            no_trigger_time = str(trigger_time).replace(
+                trigger_time[res.start() : res.end()], "."
+            )
+            os.rename(trigger_time, no_trigger_time)
 
-                nifti_file = path.join(output_path, output_filename + ".nii")
-                output_image = nifti_file + ".gz"
-
-                # Conditions to check if output NIFTI files exists,
-                # and, if DWI, if .bvec and .bval files are also present
-                nifti_exists = path.isfile(nifti_file) or path.isfile(output_image)
-                dwi_bvec_and_bval_exist = not (modality == "dwi") or (
-                    path.isfile(path.join(output_path, output_filename + ".bvec"))
-                    and path.isfile(path.join(output_path, output_filename + ".bval"))
+        # Removing images with unsupported suffixes if generated by dcm2niix
+        for suffix in ("ADC", "real", "imaginary"):
+            file_with_bad_suffix = output_path / f"{output_filename}_{suffix}"
+            if any(
+                [
+                    file_with_bad_suffix.with_suffix(s).exists()
+                    for s in (".nii", ".nii.gz")
+                ]
+            ):
+                cprint(
+                    f"Image with bad suffix {suffix} was generated by dcm2niix "
+                    f"for subject {subject} and session {session} : "
+                    f"{file_with_bad_suffix.with_suffix('.nii.gz')}. This image will NOT "
+                    "be converted as the suffix is not supported by Clinica.",
+                    lvl="warning",
                 )
+                for file_to_delete in output_path.glob(f"{output_filename}_{suffix}*"):
+                    cprint(f"Deleting file {file_to_delete}.", lvl="info")
+                    os.remove(file_to_delete)
 
-                # Check if conversion worked (output files exist)
-                if not nifti_exists or not dwi_bvec_and_bval_exist:
-                    cprint(
-                        msg=f"Conversion with dcm2niix failed for subject {subject} and session {session}",
-                        lvl="warning",
-                    )
-                    return nan
+        # Conditions to check if output NIFTI files exists,
+        # and, if DWI, if .bvec and .bval files are also present
+        nifti_exists = (
+            file_without_extension.with_suffix(".nii").is_file()
+            or output_image.is_file()
+        )
+        dwi_bvec_and_bval_exist = not (modality == "dwi") or (
+            file_without_extension.with_suffix(".bvec").is_file()
+            and file_without_extension.with_suffix(".bval").is_file()
+        )
 
-                # Case when JSON file was expected, but not generated by dcm2niix
-                elif generate_json == "y" and not os.path.exists(
-                    path.join(output_path, output_filename + ".json")
-                ):
-                    cprint(
-                        msg=f"JSON file not generated by dcm2niix for subject {subject} and session {session}",
-                        lvl="warning",
-                    )
+        # Check if conversion worked (output files exist)
+        if not nifti_exists or not dwi_bvec_and_bval_exist:
+            cprint(
+                msg=f"Conversion with dcm2niix failed for subject {subject} and session {session}",
+                lvl="warning",
+            )
+            return None
 
-                if modality_specific[modality]["to_center"]:
-                    center_nifti_origin(nifti_file, output_image)
-                    os.remove(nifti_file)
+        # Case when JSON file was expected, but not generated by dcm2niix
+        elif (
+            generate_json == "y"
+            and not file_without_extension.with_suffix(".json").exists()
+        ):
+            cprint(
+                msg=f"JSON file not generated by dcm2niix for subject {subject} and session {session}",
+                lvl="warning",
+            )
 
-            else:
-                output_image = path.join(output_path, output_filename + ".nii.gz")
-                if modality_specific[modality]["to_center"]:
-                    center_nifti_origin(image_path, output_image)
-                    if not output_image:
-                        cprint(
-                            msg=(
-                                f"For subject {subject} in session {session}, "
-                                f"an error occurred whilst recentering Nifti image: {image_path}"
-                            ),
-                            lvl="error",
-                        )
-                else:
-                    shutil.copy(image_path, output_image)
-
-        # Check if there is still the folder tmp_dcm_folder and remove it
-        remove_tmp_dmc_folder(bids_dir, image_id)
-        try:
-            print(output_image)
-        except UnboundLocalError:
-            output_image = ""
-            print(output_image)
-        return output_image
+        if modality_specific[modality]["to_center"]:
+            output_image, error_msg = center_nifti_origin(
+                file_without_extension.with_suffix(".nii"), output_image
+            )
+            if error_msg:
+                cprint(msg=error_msg, lvl="error")
+                raise ValueError(error_msg)
+            file_without_extension.with_suffix(".nii").unlink()
 
     else:
-        return nan
+        if modality_specific[modality]["to_center"]:
+            output_image, error_msg = center_nifti_origin(image_path, output_image)
+            if error_msg:
+                cprint(
+                    msg=(
+                        f"For subject {subject} in session {session}, "
+                        f"an error occurred whilst recentering Nifti image: {image_path}"
+                        f"The error is: {error_msg}"
+                    ),
+                    lvl="error",
+                )
+        else:
+            shutil.copy(image_path, output_image)
+
+    # Check if there is still the folder tmp_dcm_folder and remove it
+    remove_tmp_dmc_folder(bids_dir, image_id)
+    return output_image
 
 
 def session_label_to_viscode(session_name: str) -> str:
