@@ -20,13 +20,11 @@ def cli(
     CLINICAL_DATA_DIRECTORY respectively, to a BIDS dataset in the target BIDS_DIRECTORY.
     """
     from clinica.iotools.converters.nifd_to_bids.nifd_to_bids import convert_images
-    from clinica.utils.check_dependency import check_dcm2niix
+    from clinica.utils.check_dependency import ThirdPartySoftware, check_software
     from clinica.utils.stream import cprint
 
-    check_dcm2niix()
-
+    check_software(ThirdPartySoftware.DCM2NIIX)
     convert_images(dataset_directory, bids_directory, clinical_data_directory)
-
     cprint("Conversion to BIDS succeeded.")
 
 
