@@ -5,7 +5,7 @@ import click
 from clinica import option
 from clinica.pipelines import cli_param
 from clinica.pipelines.engine import clinica_pipeline
-from clinica.utils.atlas import T1_VOLUME_ATLASES
+from clinica.utils.atlas import T1AndPetVolumeAtlasName
 
 pipeline_name = "machinelearning-classification"
 
@@ -47,7 +47,7 @@ pipeline_name = "machinelearning-classification"
 @cli_param.option_group.option(
     "-atlas",
     "--atlas",
-    type=click.Choice(T1_VOLUME_ATLASES),
+    type=click.Choice(T1AndPetVolumeAtlasName),
     help="One of the atlases generated by t1-volume or pet-volume pipeline.",
 )
 @option.global_option_group
@@ -69,21 +69,47 @@ def cli(
 ) -> None:
     """Classification based on machine learning using scikit-learn.
 
+    Parameters
+    ----------
+    caps_directory : str
 
-        GROUP_LABEL is a string defining the group label for the current analysis, which helps you keep track of different analyses.
+    group_label : str
+        String defining the group label for the current analysis, which helps you keep track of different analyses.
 
-        The third positional argument defines the type of features for classification. It can be 'RegionBased' or 'VoxelBased'.
+    orig_input_data : str
+        Defines the type of features for classification.
+        It can be 'RegionBased' or 'VoxelBased'.
 
-        The fourth positional argument defines the studied modality ('T1w' or 'PET')
+    image_type : str
+        Defines the studied modality ('T1w' or 'PET')
 
-        The fifth positional argument defines the algorithm. It can be 'DualSVM', 'LogisticRegression' or 'RandomForest'.
+    algorithm : str
+        Defines the algorithm. It can be 'DualSVM', 'LogisticRegression' or 'RandomForest'.
 
-        The sixth positional argument defines the validation method. It can be 'RepeatedHoldOut' or 'RepeatedKFoldCV'.
+    validation : str
+        Defines the validation method. It can be 'RepeatedHoldOut' or 'RepeatedKFoldCV'.
 
-        SUBJECTS_VISITS_TSV is a TSV file containing the participant_id and the session_id columns
+    subjects_visits_tsv : str
+        TSV file containing the participant_id and the session_id columns.
 
-        DIAGNOSES_TSV is a TSV file where the diagnosis for each participant (identified by a participant ID) is reported (e.g. AD, CN). It allows the algorithm to perform the dual classification (between the two labels reported).
+    diagnoses_tsv : str
+        TSV file where the diagnosis for each participant (identified by a participant ID) is reported (e.g. AD, CN).
+        It allows the algorithm to perform the dual classification (between the two labels reported).
 
+    output_directory : str
+        The output folder path.
+
+    acq_label : str, optional
+
+    suvr_reference_region : str, optional
+
+    atlas : str, optional
+
+    n_procs : int, optional
+        The number of processes to be used by the pipeline.
+
+    Notes
+    -----
     See https://aramislab.paris.inria.fr/clinica/docs/public/latest/Pipelines/MachineLearning_Classification/
     """
     from clinica.utils.exceptions import ClinicaException
@@ -115,10 +141,11 @@ def cli(
             "Clinica will now exit."
         )
 
-    if algorithm in ["LogisticRegression", "RandomForest"]:
+    if algorithm in ("LogisticRegression", "RandomForest"):
         if orig_input_data != "RegionBased" or validation != "RepeatedHoldOut":
             raise ClinicaException(
-                "LogisticRegression or RandomForest algorithm can only work on region-based featured or RepeatedHoldOut algorithm. "
+                "LogisticRegression or RandomForest algorithm can only work "
+                "on region-based featured or RepeatedHoldOut algorithm. "
                 "Clinica will now exit."
             )
 
