@@ -1,54 +1,57 @@
-import clinica.pipelines.engine as cpe
+from typing import List
+
+from clinica.pipelines.engine import Pipeline
 
 
-class SpatialSVM(cpe.Pipeline):
+class SpatialSVM(Pipeline):
     """SpatialSVM - Prepare input data for SVM with spatial and anatomical regularization.
 
     Returns:
         A clinica pipeline object containing the SpatialSVM pipeline.
     """
 
-    def check_pipeline_parameters(self):
+    def _check_pipeline_parameters(self) -> None:
         """Check pipeline parameters."""
         from clinica.utils.group import check_group_label
 
-        # Clinica compulsory parameters
         self.parameters.setdefault("group_label", None)
         check_group_label(self.parameters["group_label"])
-
         if "orig_input_data_ml" not in self.parameters.keys():
             raise KeyError(
                 "Missing compulsory orig_input_data key in pipeline parameter."
             )
-
         # Optional parameters for inputs from pet-volume pipeline
         self.parameters.setdefault("acq_label", None)
         self.parameters.setdefault("suvr_reference_region", None)
         self.parameters.setdefault("use_pvc_data", False)
-
         # Advanced parameters
         self.parameters.setdefault("fwhm", 4)
 
-    def check_custom_dependencies(self):
+    def _check_custom_dependencies(self) -> None:
         """Check dependencies that can not be listed in the `info.json` file."""
+        pass
 
-    def get_input_fields(self):
+    def get_input_fields(self) -> List[str]:
         """Specify the list of possible inputs of this pipeline.
 
-        Returns:
+        Returns
+        -------
+        list of str :
             A list of (string) input fields name.
         """
         return ["dartel_input", "input_image"]
 
-    def get_output_fields(self):
+    def get_output_fields(self) -> List[str]:
         """Specify the list of possible outputs of this pipeline.
 
-        Returns:
+        Returns
+        -------
+        list fo str :
             A list of (string) output fields name.
         """
         return ["regularized_image"]
 
-    def build_input_node(self):
+    def _build_input_node(self):
         """Build and connect an input node to the pipeline."""
         import os
 
@@ -63,12 +66,9 @@ class SpatialSVM(cpe.Pipeline):
         from clinica.utils.inputs import clinica_file_reader, clinica_group_reader
         from clinica.utils.ux import print_groups_in_caps_directory
 
-        # Check that group already exists
-        if not os.path.exists(
-            os.path.join(
-                self.caps_directory, "groups", "group-" + self.parameters["group_label"]
-            )
-        ):
+        if not (
+            self.caps_directory / "groups" / f"group-{self.parameters['group_label']}"
+        ).exists():
             print_groups_in_caps_directory(self.caps_directory)
             raise ClinicaException(
                 f"Group {self.parameters['group_label']} does not exist. "
@@ -146,19 +146,26 @@ class SpatialSVM(cpe.Pipeline):
         read_parameters_node.inputs.dartel_input = dartel_input
         read_parameters_node.inputs.input_image = input_image
 
-        # fmt: off
         self.connect(
             [
-                (read_parameters_node, self.input_node, [("dartel_input", "dartel_input")]),
-                (read_parameters_node, self.input_node, [("input_image", "input_image")]),
+                (
+                    read_parameters_node,
+                    self.input_node,
+                    [("dartel_input", "dartel_input")],
+                ),
+                (
+                    read_parameters_node,
+                    self.input_node,
+                    [("input_image", "input_image")],
+                ),
             ]
         )
-        # fmt: on
 
-    def build_output_node(self):
+    def _build_output_node(self):
         """Build and connect an output node to the pipeline."""
+        pass
 
-    def build_core_nodes(self):
+    def _build_core_nodes(self):
         """Build and connect the core nodes of the pipeline."""
         import nipype.interfaces.io as nio
         import nipype.interfaces.utility as nutil
@@ -198,7 +205,7 @@ class SpatialSVM(cpe.Pipeline):
         heat_solver_equation.inputs.FWHM = self.parameters["fwhm"]
 
         datasink = npe.Node(nio.DataSink(), name="sinker")
-        datasink.inputs.base_directory = self.caps_directory
+        datasink.inputs.base_directory = str(self.caps_directory)
         datasink.inputs.parameterization = True
         if self.parameters["orig_input_data_ml"] == "t1-volume":
             datasink.inputs.regexp_substitutions = [
@@ -251,21 +258,49 @@ class SpatialSVM(cpe.Pipeline):
                     + r"_space-Ixi549Space_gram.npy",
                 ),
             ]
-        # Connection
-        # ==========
-        # fmt: off
         self.connect(
             [
-                (self.input_node, fisher_tensor_generation, [("dartel_input", "dartel_input")]),
-                (fisher_tensor_generation, time_step_generation, [("fisher_tensor", "g")]),
-                (self.input_node, time_step_generation, [("dartel_input", "dartel_input")]),
-                (self.input_node, heat_solver_equation, [("input_image", "input_image")]),
-                (fisher_tensor_generation, heat_solver_equation, [("fisher_tensor", "g")]),
+                (
+                    self.input_node,
+                    fisher_tensor_generation,
+                    [("dartel_input", "dartel_input")],
+                ),
+                (
+                    fisher_tensor_generation,
+                    time_step_generation,
+                    [("fisher_tensor", "g")],
+                ),
+                (
+                    self.input_node,
+                    time_step_generation,
+                    [("dartel_input", "dartel_input")],
+                ),
+                (
+                    self.input_node,
+                    heat_solver_equation,
+                    [("input_image", "input_image")],
+                ),
+                (
+                    fisher_tensor_generation,
+                    heat_solver_equation,
+                    [("fisher_tensor", "g")],
+                ),
                 (time_step_generation, heat_solver_equation, [("t_step", "t_step")]),
-                (self.input_node, heat_solver_equation, [("dartel_input", "dartel_input")]),
-                (fisher_tensor_generation, datasink, [("fisher_tensor_path", "fisher_tensor_path")]),
+                (
+                    self.input_node,
+                    heat_solver_equation,
+                    [("dartel_input", "dartel_input")],
+                ),
+                (
+                    fisher_tensor_generation,
+                    datasink,
+                    [("fisher_tensor_path", "fisher_tensor_path")],
+                ),
                 (time_step_generation, datasink, [("json_file", "json_file")]),
-                (heat_solver_equation, datasink, [("regularized_image", "regularized_image")]),
+                (
+                    heat_solver_equation,
+                    datasink,
+                    [("regularized_image", "regularized_image")],
+                ),
             ]
         )
-        # fmt: on
