@@ -1,12 +1,15 @@
+from pathlib import Path
+from typing import List
+
 from nipype import config
 
-import clinica.pipelines.engine as cpe
+from clinica.pipelines.engine import Pipeline
 
 cfg = dict(execution={"parameterize_dirs": False})
 config.update_config(cfg)
 
 
-class T1FreeSurfer(cpe.Pipeline):
+class T1FreeSurfer(Pipeline):
     """FreeSurfer-based processing of T1-weighted MR images.
 
     Returns:
@@ -14,22 +17,22 @@ class T1FreeSurfer(cpe.Pipeline):
     """
 
     @staticmethod
-    def get_processed_images(caps_directory, subjects, sessions):
-        import os
-
+    def get_processed_images(
+        caps_directory: Path, subjects: List[str], sessions: List[str]
+    ) -> List[str]:
         from clinica.utils.filemanip import extract_image_ids
         from clinica.utils.input_files import T1_FS_DESTRIEUX
         from clinica.utils.inputs import clinica_file_reader
 
-        image_ids = []
-        if os.path.isdir(caps_directory):
+        image_ids: List[str] = []
+        if caps_directory.is_dir():
             t1_freesurfer_files, _ = clinica_file_reader(
                 subjects, sessions, caps_directory, T1_FS_DESTRIEUX, False
             )
             image_ids = extract_image_ids(t1_freesurfer_files)
         return image_ids
 
-    def check_pipeline_parameters(self):
+    def _check_pipeline_parameters(self) -> None:
         """Check pipeline parameters."""
         from clinica.utils.stream import cprint
 
@@ -42,41 +45,46 @@ class T1FreeSurfer(cpe.Pipeline):
                 lvl="warning",
             )
 
-    def check_custom_dependencies(self):
+    def _check_custom_dependencies(self) -> None:
         """Check dependencies that can not be listed in the `info.json` file."""
+        pass
 
-    def get_input_fields(self):
+    def get_input_fields(self) -> List[str]:
         """Specify the list of possible inputs of this pipeline.
 
-        Note:
-            The list of inputs of the T1FreeSurfer pipeline is:
-                * t1w (str): Path of the T1 weighted image in BIDS format
+        Notes
+        -----
+        The list of inputs of the T1FreeSurfer pipeline is:
+            * t1w (str): Path of the T1 weighted image in BIDS format
 
-        Returns:
+        Returns
+        -------
+        list of str :
             A list of (string) input fields name.
         """
         return ["t1w"]
 
-    def get_output_fields(self):
+    def get_output_fields(self) -> List[str]:
         """Specify the list of possible outputs of this pipeline.
 
-        Note:
-            The list of outputs of the T1FreeSurfer pipeline is:
-                * image_id (str): Image ID (e.g. sub-CLNC01_ses-M000)
+        Notes
+        -----
+        The list of outputs of the T1FreeSurfer pipeline is:
+            * image_id (str): Image ID (e.g. sub-CLNC01_ses-M000)
 
-        Returns:
+        Returns
+        -------
+        list of str :
             A list of (string) output fields name.
         """
         return ["image_id"]
 
-    def build_input_node(self):
+    def _build_input_node(self):
         """Build and connect an input node to the pipeline.
 
         Raise:
             ClinicaBIDSError: If there are duplicated files or missing files for any subject
         """
-        import os
-
         import nipype.interfaces.utility as nutil
         import nipype.pipeline.engine as npe
 
@@ -117,11 +125,9 @@ class T1FreeSurfer(cpe.Pipeline):
             else:
                 cprint(msg="Image(s) will be ignored by Clinica.", lvl="warning")
                 input_ids = [
-                    p_id + "_" + s_id
-                    for p_id, s_id in zip(self.subjects, self.sessions)
+                    f"{p_id}_{s_id}" for p_id, s_id in zip(self.subjects, self.sessions)
                 ]
                 to_process_ids = list(set(input_ids) - set(processed_ids))
-
                 self.subjects, self.sessions = extract_subjects_sessions_from_filename(
                     to_process_ids
                 )
@@ -143,17 +149,13 @@ class T1FreeSurfer(cpe.Pipeline):
         if not t1w_files:
             raise ClinicaException("Empty dataset or already processed")
 
-        # Save subjects to process in <WD>/<Pipeline.name>/participants.tsv
-        folder_participants_tsv = os.path.join(self.base_dir, self.name)
         save_participants_sessions(
-            self.subjects, self.sessions, folder_participants_tsv
+            self.subjects, self.sessions, self.base_dir / self.name
         )
-
         if len(self.subjects):
             print_images_to_process(self.subjects, self.sessions)
             cprint(
-                "List available in %s"
-                % os.path.join(folder_participants_tsv, "participants.tsv")
+                f"List available in {self.base_dir / self.name / 'participants.tsv'}"
             )
             cprint("The pipeline will last approximately 10 hours per image.")
 
@@ -177,10 +179,8 @@ class T1FreeSurfer(cpe.Pipeline):
             ]
         )
 
-    def build_output_node(self):
+    def _build_output_node(self):
         """Build and connect an output node to the pipeline."""
-        import os
-
         import nipype.interfaces.utility as nutil
         import nipype.pipeline.engine as npe
 
@@ -194,9 +194,7 @@ class T1FreeSurfer(cpe.Pipeline):
             ),
             name="SaveToCaps",
         )
-        save_to_caps.inputs.source_dir = os.path.join(
-            self.base_dir, self.name, "ReconAll"
-        )
+        save_to_caps.inputs.source_dir = self.base_dir / self.name / "ReconAll"
         save_to_caps.inputs.caps_dir = self.caps_directory
         save_to_caps.inputs.overwrite_caps = self.overwrite_caps
 
@@ -206,10 +204,8 @@ class T1FreeSurfer(cpe.Pipeline):
             ]
         )
 
-    def build_core_nodes(self):
+    def _build_core_nodes(self):
         """Build and connect the core nodes of the pipeline."""
-        import os
-
         import nipype.interfaces.utility as nutil
         import nipype.pipeline.engine as npe
         from nipype.interfaces.freesurfer.preprocess import ReconAll
@@ -232,9 +228,7 @@ class T1FreeSurfer(cpe.Pipeline):
             name="0-InitPipeline",
         )
         init_input.inputs.recon_all_args = self.parameters["recon_all_args"]
-        init_input.inputs.output_dir = os.path.join(
-            self.base_dir, self.name, "ReconAll"
-        )
+        init_input.inputs.output_dir = self.base_dir / self.name / "ReconAll"
 
         # Run recon-all command
         # FreeSurfer segmentation will be in <subjects_dir>/<image_id>/
