@@ -11,11 +11,10 @@ def atlas_statistics(in_image, atlas_list):
     Returns:
         List of paths to TSV files
     """
-    from os.path import abspath, join
+    from pathlib import Path
 
     from nipype.utils.filemanip import split_filename
 
-    from clinica.utils.atlas import AtlasAbstract
     from clinica.utils.filemanip import get_subject_id
     from clinica.utils.statistics import statistics_on_atlas
     from clinica.utils.ux import print_end_image
@@ -23,15 +22,12 @@ def atlas_statistics(in_image, atlas_list):
     subject_id = get_subject_id(in_image)
 
     orig_dir, base, ext = split_filename(in_image)
-    atlas_classes = AtlasAbstract.__subclasses__()
     atlas_statistics_list = []
     for atlas in atlas_list:
-        for atlas_class in atlas_classes:
-            if atlas_class.get_name_atlas() == atlas:
-                out_atlas_statistics = abspath(
-                    join(f"./{base}_space-{atlas}_map-graymatter_statistics.tsv")
-                )
-                statistics_on_atlas(in_image, atlas_class(), out_atlas_statistics)
-                atlas_statistics_list.append(out_atlas_statistics)
+        out_atlas_statistics = Path(
+            f"./{base}_space-{atlas}_map-graymatter_statistics.tsv"
+        ).resolve()
+        statistics_on_atlas(in_image, atlas, out_atlas_statistics)
+        atlas_statistics_list.append(out_atlas_statistics)
     print_end_image(subject_id)
     return atlas_statistics_list
