@@ -4,10 +4,9 @@ from nipype.interfaces.base import File, TraitedSpec, traits
 from nipype.interfaces.spm.base import SPMCommand, SPMCommandInputSpec, scans_for_fnames
 from nipype.utils.filemanip import split_filename
 
-from clinica.utils.spm import spm_standalone_is_available, use_spm_standalone
+from clinica.utils.spm import use_spm_standalone_if_available
 
-if spm_standalone_is_available():
-    use_spm_standalone()
+use_spm_standalone_if_available()
 
 
 class DARTELExistingTemplateInputSpec(SPMCommandInputSpec):
