@@ -1,24 +1,25 @@
-# Use hash instead of parameters for iterables folder names
-# Otherwise path will be too long and generate OSError
+from typing import List
+
 from nipype import config
 
-import clinica.pipelines.engine as cpe
+from clinica.pipelines.engine import Pipeline
 
 cfg = dict(execution={"parameterize_dirs": False})
 config.update_config(cfg)
 
 
-class T1VolumeTissueSegmentation(cpe.Pipeline):
+class T1VolumeTissueSegmentation(Pipeline):
     """T1VolumeTissueSegmentation - Tissue segmentation, bias correction and spatial normalization to MNI space.
 
     Returns:
         A clinica pipeline object containing the T1VolumeTissueSegmentation pipeline.
     """
 
-    def check_custom_dependencies(self):
+    def _check_custom_dependencies(self) -> None:
         """Check dependencies that can not be listed in the `info.json` file."""
+        pass
 
-    def check_pipeline_parameters(self):
+    def _check_pipeline_parameters(self) -> None:
         """Check pipeline parameters."""
         from clinica.utils.spm import get_tpm
 
@@ -27,23 +28,26 @@ class T1VolumeTissueSegmentation(cpe.Pipeline):
         self.parameters.setdefault("save_warped_unmodulated", True)
         self.parameters.setdefault("save_warped_modulated", False)
         self.parameters.setdefault("tissue_probability_maps", None)
-
         # Fetch the TPM in case none was passed explicitly.
         if not self.parameters["tissue_probability_maps"]:
             self.parameters["tissue_probability_maps"] = get_tpm()
 
-    def get_input_fields(self):
+    def get_input_fields(self) -> List[str]:
         """Specify the list of possible inputs of this pipeline.
 
-        Returns:
+        Returns
+        -------
+        list of str :
             A list of (string) input fields name.
         """
         return ["t1w"]
 
-    def get_output_fields(self):
+    def get_output_fields(self) -> List[str]:
         """Specify the list of possible outputs of this pipeline.
 
-        Returns:
+        Returns
+        -------
+        list of str :
             A list of (string) output fields name.
         """
         return [
@@ -59,7 +63,7 @@ class T1VolumeTissueSegmentation(cpe.Pipeline):
             "t1_mni",
         ]
 
-    def build_input_node(self):
+    def _build_input_node(self):
         """Build and connect an input node to the pipeline.
 
         Raise:
@@ -112,10 +116,11 @@ class T1VolumeTissueSegmentation(cpe.Pipeline):
             ]
         )
 
-    def build_output_node(self):
+    def _build_output_node(self):
         """Build and connect an output node to the pipeline."""
+        pass
 
-    def build_core_nodes(self):
+    def _build_core_nodes(self):
         """Build and connect the core nodes of the pipeline."""
         import nipype.interfaces.io as nio
         import nipype.interfaces.spm as spm
@@ -124,7 +129,7 @@ class T1VolumeTissueSegmentation(cpe.Pipeline):
 
         from clinica.utils.filemanip import unzip_nii, zip_nii
         from clinica.utils.nipype import container_from_filename, fix_join
-        from clinica.utils.spm import spm_standalone_is_available, use_spm_standalone
+        from clinica.utils.spm import use_spm_standalone_if_available
 
         from .t1_volume_tissue_segmentation_utils import (
             ApplySegmentationDeformation,
@@ -134,8 +139,7 @@ class T1VolumeTissueSegmentation(cpe.Pipeline):
             zip_list_files,
         )
 
-        if spm_standalone_is_available():
-            use_spm_standalone()
+        use_spm_standalone_if_available()
 
         # Get <subject_id> (e.g. sub-CLNC01_ses-M000) from input_node
         # and print begin message
@@ -149,8 +153,6 @@ class T1VolumeTissueSegmentation(cpe.Pipeline):
             name="0-InitNode",
         )
 
-        # Unzipping
-        # =========
         unzip_node = npe.Node(
             nutil.Function(
                 input_names=["in_file"], output_names=["out_file"], function=unzip_nii
@@ -183,10 +185,6 @@ class T1VolumeTissueSegmentation(cpe.Pipeline):
             ),
             name="WriteEndMessage",
         )
-
-        # Connection
-        # ==========
-        # fmt: off
         self.connect(
             [
                 (self.input_node, init_node, [("t1w", "t1w")]),
@@ -194,21 +192,30 @@ class T1VolumeTissueSegmentation(cpe.Pipeline):
                 (unzip_node, new_segment, [("out_file", "channel_files")]),
                 (init_node, print_end_message, [("subject_id", "subject_id")]),
                 (unzip_node, t1_to_mni, [("out_file", "in_files")]),
-                (new_segment, t1_to_mni, [("forward_deformation_field", "deformation_field")]),
-                (new_segment, self.output_node, [("bias_corrected_images", "bias_corrected_images"),
-                                                 ("bias_field_images", "bias_field_images"),
-                                                 ("dartel_input_images", "dartel_input_images"),
-                                                 ("forward_deformation_field", "forward_deformation_field"),
-                                                 ("inverse_deformation_field", "inverse_deformation_field"),
-                                                 ("modulated_class_images", "modulated_class_images"),
-                                                 ("native_class_images", "native_class_images"),
-                                                 ("normalized_class_images", "normalized_class_images"),
-                                                 ("transformation_mat", "transformation_mat")]),
+                (
+                    new_segment,
+                    t1_to_mni,
+                    [("forward_deformation_field", "deformation_field")],
+                ),
+                (
+                    new_segment,
+                    self.output_node,
+                    [
+                        ("bias_corrected_images", "bias_corrected_images"),
+                        ("bias_field_images", "bias_field_images"),
+                        ("dartel_input_images", "dartel_input_images"),
+                        ("forward_deformation_field", "forward_deformation_field"),
+                        ("inverse_deformation_field", "inverse_deformation_field"),
+                        ("modulated_class_images", "modulated_class_images"),
+                        ("native_class_images", "native_class_images"),
+                        ("normalized_class_images", "normalized_class_images"),
+                        ("transformation_mat", "transformation_mat"),
+                    ],
+                ),
                 (t1_to_mni, self.output_node, [("out_files", "t1_mni")]),
                 (self.output_node, print_end_message, [("t1_mni", "final_file")]),
             ]
         )
-        # fmt: on
 
         # Find container path from t1w filename
         # =====================================
@@ -224,7 +231,7 @@ class T1VolumeTissueSegmentation(cpe.Pipeline):
         # Writing CAPS
         # ============
         write_node = npe.Node(name="WriteCAPS", interface=nio.DataSink())
-        write_node.inputs.base_directory = self.caps_directory
+        write_node.inputs.base_directory = str(self.caps_directory)
         write_node.inputs.parameterization = False
         write_node.inputs.regexp_substitutions = [
             (r"(.*)c1(sub-.*)(\.nii(\.gz)?)$", r"\1\2_segm-graymatter\3"),
@@ -255,28 +262,77 @@ class T1VolumeTissueSegmentation(cpe.Pipeline):
             (r"trait_added", r""),
         ]
 
-        # fmt: off
         self.connect(
             [
                 (self.input_node, container_path, [("t1w", "bids_or_caps_filename")]),
-                (container_path, write_node, [(("container", fix_join, "t1", "spm", "segmentation"), "container")]),
-                (self.output_node, write_node, [(("native_class_images", zip_list_files, True), "native_space"),
-                                                (("dartel_input_images", zip_list_files, True), "dartel_input")]),
-                (self.output_node, write_node, [(("inverse_deformation_field", zip_nii, True), "inverse_deformation_field")]),
-                (self.output_node, write_node, [(("forward_deformation_field", zip_nii, True), "forward_deformation_field")]),
+                (
+                    container_path,
+                    write_node,
+                    [
+                        (
+                            ("container", fix_join, "t1", "spm", "segmentation"),
+                            "container",
+                        )
+                    ],
+                ),
+                (
+                    self.output_node,
+                    write_node,
+                    [
+                        (("native_class_images", zip_list_files, True), "native_space"),
+                        (("dartel_input_images", zip_list_files, True), "dartel_input"),
+                    ],
+                ),
+                (
+                    self.output_node,
+                    write_node,
+                    [
+                        (
+                            ("inverse_deformation_field", zip_nii, True),
+                            "inverse_deformation_field",
+                        )
+                    ],
+                ),
+                (
+                    self.output_node,
+                    write_node,
+                    [
+                        (
+                            ("forward_deformation_field", zip_nii, True),
+                            "forward_deformation_field",
+                        )
+                    ],
+                ),
                 (self.output_node, write_node, [(("t1_mni", zip_nii, True), "t1_mni")]),
             ]
         )
         if self.parameters["save_warped_unmodulated"]:
             self.connect(
                 [
-                    (self.output_node, write_node, [(("normalized_class_images", zip_list_files, True), "normalized")]),
+                    (
+                        self.output_node,
+                        write_node,
+                        [
+                            (
+                                ("normalized_class_images", zip_list_files, True),
+                                "normalized",
+                            )
+                        ],
+                    ),
                 ]
             )
         if self.parameters["save_warped_modulated"]:
             self.connect(
                 [
-                    (self.output_node, write_node, [(("modulated_class_images", zip_list_files, True), "modulated_normalized")]),
+                    (
+                        self.output_node,
+                        write_node,
+                        [
+                            (
+                                ("modulated_class_images", zip_list_files, True),
+                                "modulated_normalized",
+                            )
+                        ],
+                    ),
                 ]
             )
-        # fmt: on
