@@ -88,16 +88,14 @@ def build_core_workflow(name: str = "core", parameters={}) -> Workflow:
     wf : Workflow
         The core workflow.
     """
-    from os.path import dirname, join
+    from os.path import join
     from typing import Any
 
     import clinica.pydra.statistics_volume.task as utils
-    from clinica.utils.exceptions import ClinicaException
     from clinica.utils.filemanip import get_parent
-    from clinica.utils.spm import spm_standalone_is_available, use_spm_standalone
+    from clinica.utils.spm import use_spm_standalone_if_available
 
-    if spm_standalone_is_available():
-        use_spm_standalone()
+    use_spm_standalone_if_available()
 
     parameters = _check_pipeline_parameters(parameters)
 
