@@ -1,3 +1,5 @@
+from pathlib import Path
+
 import pytest
 
 from clinica.pydra.query import BIDSQuery, CAPSFileQuery, CAPSGroupQuery, Query
@@ -39,12 +41,20 @@ def test_caps_file_query():
     assert q.query == {
         "mask_tissues": [
             {
-                "pattern": "t1/spm/segmentation/normalized_space/*_*_T1w_segm-graymatter_space-Ixi549Space_modulated-off_probability.nii*",
+                "pattern": Path("t1")
+                / "spm"
+                / "segmentation"
+                / "normalized_space"
+                / "*_*_T1w_segm-graymatter_space-Ixi549Space_modulated-off_probability.nii*",
                 "description": "Tissue probability map graymatter based on native MRI in MNI space (Ixi549) without modulation.",
                 "needed_pipeline": "t1-volume-tissue-segmentation",
             },
             {
-                "pattern": "t1/spm/segmentation/normalized_space/*_*_T1w_segm-whitematter_space-Ixi549Space_modulated-off_probability.nii*",
+                "pattern": Path("t1")
+                / "spm"
+                / "segmentation"
+                / "normalized_space"
+                / "*_*_T1w_segm-whitematter_space-Ixi549Space_modulated-off_probability.nii*",
                 "description": "Tissue probability map whitematter based on native MRI in MNI space (Ixi549) without modulation.",
                 "needed_pipeline": "t1-volume-tissue-segmentation",
             },
@@ -61,18 +71,30 @@ def test_caps_file_query():
     assert q.query == {
         "mask_tissues": [
             {
-                "pattern": "t1/spm/segmentation/normalized_space/*_*_T1w_segm-graymatter_space-Ixi549Space_modulated-off_probability.nii*",
+                "pattern": Path("t1")
+                / "spm"
+                / "segmentation"
+                / "normalized_space"
+                / "*_*_T1w_segm-graymatter_space-Ixi549Space_modulated-off_probability.nii*",
                 "description": "Tissue probability map graymatter based on native MRI in MNI space (Ixi549) without modulation.",
                 "needed_pipeline": "t1-volume-tissue-segmentation",
             },
             {
-                "pattern": "t1/spm/segmentation/normalized_space/*_*_T1w_segm-whitematter_space-Ixi549Space_modulated-off_probability.nii*",
+                "pattern": Path("t1")
+                / "spm"
+                / "segmentation"
+                / "normalized_space"
+                / "*_*_T1w_segm-whitematter_space-Ixi549Space_modulated-off_probability.nii*",
                 "description": "Tissue probability map whitematter based on native MRI in MNI space (Ixi549) without modulation.",
                 "needed_pipeline": "t1-volume-tissue-segmentation",
             },
         ],
         "flow_fields": {
-            "pattern": "t1/spm/dartel/group-UnitTest/sub-*_ses-*_T1w_target-UnitTest_transformation-forward_deformation.nii*",
+            "pattern": Path("t1")
+            / "spm"
+            / "dartel"
+            / "group-UnitTest"
+            / "sub-*_ses-*_T1w_target-UnitTest_transformation-forward_deformation.nii*",
             "description": "Deformation from native space to group template UnitTest space.",
             "needed_pipeline": "t1-volume-create-dartel",
         },
@@ -84,7 +106,7 @@ def test_caps_group_query():
     assert len(q) == 1
     assert q.query == {
         "dartel_template": {
-            "pattern": "group-UnitTest/t1/group-UnitTest_template.nii*",
+            "pattern": Path("group-UnitTest") / "t1" / "group-UnitTest_template.nii*",
             "description": "T1w template file of group UnitTest",
             "needed_pipeline": "t1-volume or t1-volume-create-dartel",
         }
